+++ paper.bib
@@ -4,6 +4,7 @@
   journal={Computational Statistics},
   volume=2,
   pages={317-328},
+  doi = {10.1007/s00180-009-0178-4},
   year=2011
 }
 
@@ -11,6 +12,7 @@
   title={Crime and place: a longitudinal examination of street segment patterns in Vancouver, BC},
   author={Curman, A.S.N. and Andresen, M.A. and Brantingham, P.J.},
   journal={Journal of Quantitative Criminology},
+  doi={10.1007/s10940-014-9228-3},
   volume=31,
   pages={127-147},
   year=2015
@@ -47,6 +49,7 @@
   title={Communities, street guns and homicide trajectories in Chicago, 1980-1995: merging methods for examining homicide trends across space and time},
   author={Griffiths, E. and Chavez, J.M.},
   journal={Criminology},
+  doi = {10.1111/j.1745-9125.2004.tb00541.x},
   volume=42,
   pages={941-978},
   year=2004
@@ -56,6 +59,7 @@
   title={The trajectories of crime at places: understanding the patterns of disaggregated crime types},
   author={Andresen, M.A. and Curman, A.S. and Linning, S.J.},
   journal={Journal of Quantitative Criminology},
+  doi = {10.1007/s10940-016-9301-1},
   volume=33,
   pages={427-449},
   year=2017
@@ -65,6 +69,7 @@
   title={Trajectories of Crime at Places: A Longitudinal Study of Street Segments in the City of Seattle},
   author={Weisburd, D. and Bushway, S. and Lum, C. and Yang, S.},
   journal={Criminology},
+  doi = {10.4324/9781315089256-19},
   volume=42,
   pages={283-322},
   year=2004
@@ -74,6 +79,7 @@
   title={Neighborhood dynamics of urban violence: Understanding the immigration connection},
   author={Chavez, J.M. and Griffiths, E.},
   journal={Homicide Studies},
+  doi = {10.1177/1088767909337701}
   volume=13,
   pages={261-273},
   year=2009
@@ -83,6 +89,7 @@
   title={Local variance in the crime drop: A longitudinal study of neighbourhoods in greater Glasgow, Scotland},
   author={Bannister, J. and Bates, E. and Kearns, A.},
   journal={British Journal of Criminology},
+  doi = {10.1093/bjc/azx022},
   volume=58,
   pages={177-199},
   year=2017
@@ -92,6 +99,7 @@
   title={A classification EM algorithm for clustering and two stochastic versions},
   author={Celeux, G. and Govaert, G.},
   journal={Journal of Computational Statistics and Data Analysis},
+  doi = {10.1016/0167-9473(92)90042-e},
   volume=14,
   pages={315-332},
   year=1992
@@ -101,6 +109,7 @@
   title={A dendrite method for cluster analysis},
   author={Calinski, T. and Harabasz, J.},
   journal={Communications in Statistics-theory and Methods},
+  doi = {10.1080/03610917408548446},
   volume=3,
   pages={1-27},
   year=1974
@@ -110,6 +119,7 @@
   title={Silhouettes: A graphical aid to the interpretation and validation of cluster analysis},
   author={Rousseeuw, P.J.},
   journal={Journal of Computational and Applied Mathematics},
+  doi = {10.1016/0377-0427(87)90125-7},
   pages={53-65},
   year=1987
 }
