+++ paper.bib
@@ -79,7 +79,7 @@
   title={Neighborhood dynamics of urban violence: Understanding the immigration connection},
   author={Chavez, J.M. and Griffiths, E.},
   journal={Homicide Studies},
-  doi = {10.1177/1088767909337701}
+  doi = {10.1177/1088767909337701},
   volume=13,
   pages={261-273},
   year=2009
