### 2020-01-17 #2019 @MAnalytics
Missing DOIs now included!
Thanks you.
