### 2020-01-22 #2019 @MAnalytics
Dear Sir,

Below are my suggestions (based on this list of people<https://bit.ly/joss-reviewers>) for potential reviewers of our article:

1.       rcannood

2.       sjsrey

3.       seabbs

4.       Maczokni

I hope to hear from you soon.

Monsuru.
