### 2020-01-23 #2019 @karthik
👋 @MAnalytics I'l be the editor handling your submission. A quick look shows that you have no unit tests at all. Can you please add sufficient tests to cover all functionality of your package?
