### 2020-01-26 6ea7731
test manip. functions
