+++ test/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(akmedoids)
+test_check("akmedoids")
+
