+++ DESCRIPTION
@@ -16,7 +16,6 @@ RoxygenNote: 7.0.1
 Suggests:
     knitr,
     rmarkdown,
-    flextable,
     kableExtra,
     clusterCrit,
     testthat,
