### 2020-01-27 fc59354
vignette, remove flextable
