+++ R/alphaLabel.R
@@ -11,6 +11,7 @@
 #' @export
 
 alphaLabel <- function(x){
+
   combind_A <- LETTERS
   combind <-  combn(LETTERS, m=2, sep="")# combind[1:2,]
   list_Letters <- NULL
