+++ test/testthat/test-clustering.R
@@ -1,0 +1,6 @@
+context("clustering test")
+
+df <- data.frame(id=c("u","v","w","x","y"),
+                 a=c(2, 3, 5, 2, 0),
+                 b=c(5, 1, 0, 3, 1),
+                 c=c(5, 1, 4, 2, 7))
