+++ test/testthat/test-stats.R
@@ -1,0 +1,2 @@
+
+context("testing statistics and plot")
