+++ test/testthat/test-stats.R
@@ -1,2 +1,0 @@
-
-context("testing statistics and plot")
