### 2020-01-28 dfe9811
Update statPrint.Rd
