### 2020-01-28 04deb5f
update clustering fct
