+++ test/testthat/test-clustering.R
@@ -27,13 +27,36 @@ test_that("test clustr procedures and the output specification", {
 })
 
 
-clustr <- akmedoids.clust(df, id_field = TRUE,
-                          k = 3)
-clustr <- as.vector(clustr$memberships)
-print(statPrint(clustr, df, id_field=TRUE, type="lines", y.scaling="fixed"))
+clust1 <- as.vector(clust1$memberships)
+stats = statPrint(clust1, df, id_field=TRUE,
+          type="lines", y.scaling="fixed", showplots=FALSE)
+
+gr1 <- as.numeric(as.character(stats$descriptiveStats$n))[1]
+gr2 <- as.numeric(as.character(stats$descriptiveStats$n))[2]
+gr3 <- as.numeric(as.character(stats$descriptiveStats$n))[3]
+
+#test increment of clust
+incr = function(a, b){
+  if(a>=b){
+    check = 1}
+  else{
+    check = 2
+  }
+  return(check)
+}
+
+st1 <- as.numeric(as.character(stats$changeStats$`%+ve Traj.`))[1]
+st2 <- as.numeric(as.character(stats$changeStats$`%+ve Traj.`))[2]
+st3 <- as.numeric(as.character(stats$changeStats$`%+ve Traj.`))[3]
 
 test_that("test cluster stats routine is okay", {
+    expect_equal(gr1, 3)
+    expect_equal(gr2, 1)
+    expect_equal(gr3, 1)
 
+    expect_equal(incr(st2, st1), 1)
+    expect_equal(incr(st3, st1), 1)
+    expect_equal(incr(st3, st2), 1)
 
 })
 
