### 2020-02-04 #2019 @karthik
Thanks @MAnalytics
I'm looking for reviewers now. Other things you might want to do in the meantime:

- Add some installation instructions to your README
- omit "Date" in DESCRIPTION. It is not required and it
    gets invalid quite often. A build date will be added to the package
    when you perform `R CMD build` on it.
- add a "URL" field to DESCRIPTION. It helps users find
    information about your package online. If your package does not
    have a homepage, add an URL to GitHub, or the CRAN package package
    page.
- add a "BugReports" field to DESCRIPTION, and point it to
    a bug tracker. Many online code hosting services provide bug
    trackers for free, https://github.com, https://gitlab.com, etc.

For the last two, see [example here](https://github.com/karthik/holepunch/blob/master/DESCRIPTION#L17-L18)
