### 2020-02-04 #2019 @karthik
A couple more comments:

- avoid long code lines, it is bad for readability. Also,

Use the Rstudio code formatting option under edit to break long lines of code.

Also

- avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
    result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

These appear mostly in `clust.R`.
