+++ README.md
@@ -6,6 +6,8 @@ An R package for analysing and clustering longitudinal data
 
 The `akmedoids` package advances a set of R-functions for longitudinal clustering of long-term trajectories and determines the optimal solution based on either the `average silhouette width` (Rousseeuw P. J. (1987) or the `Caliński-Harabatz` criterion (Caliński and Harabasz 1974). The package also includes a set of functions for addressing common data issues, such as missing entries and outliers, prior to conducting advance longitudinal data analysis. One of the key objectives of this package is to facilitate easy replication of a recent paper which examined small area inequality in the crime drop (see Adepeju et al. 2019). Many of the functions provided in the `akmedoids` package may be applied to longitudinal data in general. 
 
+#For more information and usability, check out details on [CRAN](https://cran.r-project.org/web/packages/akmedoids/index.html).
+
 ### General 
 
 Many of the functions provided in the `akmedoids` package may be applied to longitudinal data in general. 
