+++ NAMESPACE
@@ -1,5 +1,14 @@
 # Generated by roxygen2: do not edit by hand
 
+export(akmedoids.clust)
+export(alphaLabel)
+export(dataImputation)
+export(elbowPoint)
+export(outlierDetect)
+export(props)
+export(rates)
+export(statPrint)
+export(wSpaces)
 import(kml, grDevices, reshape2, Hmisc, stats, utils, ggplot2, longitudinalData)
 import(reshape2, ggplot2, stats)
 import(stats)
