+++ R/akmedoids.clust.R
@@ -24,13 +24,14 @@
 #' @export
 
 
-akmedoids.clust <- function(traj, id_field = FALSE, method = "linear", k = c(3,6), crit = "Silhouette"){ #TRUE #k<-3
-
-  #traj = crime_per_000_people
+akmedoids.clust <- function(traj, id_field = FALSE,
+                            method = "linear", k = c(3,6),
+                            crit = "Silhouette"){ #TRUE #k<-3
 
 qualityCrit <- 0
 
-#create a vector if an interger of k if provided.
+#create a vector of two elements,
+#if k is a single value
   if(length(k)==1){
     k <- rep(k, 2)
   }
@@ -38,11 +39,12 @@ qualityCrit <- 0
 #linear medoid method
 if(method=="linear"){
 
-  #check if unaccepted value of k in inputted
+  #expand k
   k_ <- k[1]:k[2]
-  #checks
+
+  #check if unacceptable value of k in inputted
   if(k[1] < 3 | k[1] > nrow(traj) | k[2] > nrow(traj) |
-     k[1] > 20 | k[2] > 20 | k[1]>k[2]){
+     k[1] > 20 | k[2] > 20 | k[1]> k[2]){
     flush.console()
     print("*******Error!********")
     if(k[1] < 3 | k[1] > nrow(traj) | k[2] > nrow(traj) |
@@ -57,104 +59,102 @@ if(method=="linear"){
     stop("(: Program terminated!!! :)")
   } else {
 
-    dat <- traj
-    #check if there is id_field
-    #check if there is unique(id) field
-    if(id_field==TRUE){
-      n_CL <- colnames(dat)[1]
-      col_names <- as.vector(dat[,1])
-      dat <- dat[,2:ncol(dat)]
-      #check if the 'id_field' is a unique field
-      if(!length(col_names)==length(unique(col_names))){
-        stop("(: The 'id_field' is not a unique field. Function terminated!!! :)")
-      }
+  dat <- traj
+
+  #check if there is id_field
+  #check if there is unique (id) field
+  if(id_field==TRUE){
+    n_CL <- colnames(dat)[1]
+    col_names <- as.vector(dat[,1])
+    dat <- dat[,2:ncol(dat)]
+    #check if the 'id_field' is a unique field
+    if(!length(col_names)==length(unique(col_names))){
+      stop("(: The 'id_field' does not contain unique elements. Function terminated!!! :)")
+    }
     }
 
+  #get the 'time' vector
+  time = 1:ncol(dat)
+  #-----------------------------------------------------
+  #get the linear coefficients
+  sl_List <- NULL
+  time <- as.numeric(1:ncol(dat))
+  for(i in seq_len(nrow(dat))){ #i<-1
+    b=coefficients(lm(as.numeric(as.character(dat[i,]))~
+                          as.numeric(as.character(time))))
+    sl_List <- rbind(sl_List, cbind(as.numeric(b[1]),
+                                      as.numeric(b[2])))
+  }
 
-    #create a time list
-    time = 1:ncol(dat) #clean.time   #mode(time)
-    #-----------------------------------------------------
-    #looping through to calculate the coefficients
-    sl_List <- NULL
-    time <- as.numeric(1:ncol(dat))
-    for(i in 1:nrow(dat)){ #i<-1
-      b=coefficients(lm(as.numeric(as.character(dat[i,]))~as.numeric(as.character(time))))
-      sl_List <- rbind(sl_List, cbind(as.numeric(b[1]), as.numeric(b[2])))
+  sl_List <- as.data.frame(cbind(1:nrow(sl_List), sl_List))
+  colnames(sl_List) <- c("sn", "intersect","slope")
+
[…]
