### 2020-02-07 3a176ee
main clustering function
