### 2020-02-07 9f21281
update wspace function
