### 2020-02-07 0d7f790
add showplots argument
