+++ R/statPrint.R
@@ -27,249 +27,290 @@
 #' @references \code{Wickham H. (2016). Elegant graphics for Data Analysis. Spring-Verlag New York (2016)}
 #' @export
 
-statPrint <- function(clustr, traj, id_field=TRUE, reference = 1, N.quant = 4, showplots=TRUE, type = "lines", y.scaling="fixed"){
+statPrint <- function(clustr, traj, id_field=TRUE,
+                      reference = 1, N.quant = 4,
+                      showplots=TRUE, type = "lines",
+                      y.scaling="fixed"){
 
-  #joining the data with the clusters
-  clustr <- data.frame(cbind(traj, clusters=clustr))
+#joining the data with the clusters
+clustr <- data.frame(cbind(traj, clusters=clustr))
 
-  dat <- traj #backing up the data
-  N.quant <- round(N.quant, digits = 0)
+dat <- traj #backing up the data
+N.quant <- round(N.quant, digits = 0)
 
-  if(N.quant < 2 | N.quant > 10){
-    stop("*-------Please, enter an integer between 2 and 10 for the 'N.quant' argument'!!!-------*")
+if(N.quant < 2 | N.quant > 10){
+  stop("*-------Please, enter an integer between 2 and 10 for the 'N.quant' argument'!!!-------*")
+}
+
+#check if there is id_field
+#check if id field  is unique
+if(id_field==TRUE){
+  dat <- dat[,2:ncol(dat)]
+  n_CL <- colnames(clustr)[1]
+  col_names <- as.vector(clustr[,1])
+  if(!length(col_names)==length(unique(col_names))){
+    stop("(: The 'id_field' is not a unique field. Function terminated!!! :)")
   }
+}
 
-  #check if 'id_field' is a unique field
-  if(id_field==TRUE){
-    dat <- dat[,2:ncol(dat)]
-    n_CL <- colnames(clustr)[1]
-    col_names <- as.vector(clustr[,1])
-    if(!length(col_names)==length(unique(col_names))){
-      stop("(: The 'id_field' is not a unique field. Function terminated!!! :)")
-    }
+if(id_field==FALSE){
+  clustr <- cbind(1:nrow(clustr), clustr)
+}
+
+#cluster list
+clusters <- as.vector(clustr[,ncol(clustr)])
+data_subset <- clustr[,1:(ncol(clustr)-1)]
+data_subset <- as.data.frame(data_subset)
+colnames(data_subset) <- c("code", 1:(ncol(data_subset)-1))
+data.subset.melted <- suppressWarnings(melt(data_subset, id="code"))
+#append cluster list
+data.subset.melted <- cbind(data.subset.melted,
+                            rep(clusters, ncol(data_subset)-1))
+colnames(data.subset.melted) <- c("id","Year","value", "clusters")
+
+#----------------------------------------------------
+#preparing the data to generate descriptive statitics
+year_uni <- as.vector(unique(data.subset.melted$Year))
+order_Cluster <- as.vector(unique(data.subset.melted$clusters))
+clusters_uni <- order_Cluster[order(as.vector(unique(data.subset.melted$clusters)))]
+
+change_ave_yr_ALL <- NULL
+for(q in seq_len(length(clusters_uni))){
+  all_clust_list <-
+    data.subset.melted[which(data.subset.melted$clusters==clusters_uni[q]),]
+  ave_yr <- NULL
+  for(m in seq_len(length(year_uni))){
+    yr_ <-
+      all_clust_list[which(as.vector(all_clust_list$Year)==year_uni[m]),]
+    ave_yr <- c(ave_yr, sum(yr_$value))
   }
+  change_ave_yr_ALL <- rbind(change_ave_yr_ALL,  ave_yr)
+}
 
[…]
