### 2020-02-07 4a4f107
update removeRowsN function
