### 2020-02-07 9d5de02
update props.R
