### 2020-02-07 ff1b3f0
update OutlierDet. function
