+++ R/outlierDetect.R
@@ -23,162 +23,167 @@
 #' @return A dataframe with outlier observations replaced or removed.
 #' @export
 
-outlierDetect <- function(traj, id_field = FALSE, method = 1, threshold = 0.95, count = 1, replace_with = 1){
+outlierDetect <- function(traj, id_field = FALSE,
+                          method = 1, threshold = 0.95,
+                          count = 1, replace_with = 1){
 
-  dat <- traj
+dat <- traj
 
-  #back up data
-  b_dat <- dat    #when joining back... remember if the option is to remove some #then the crime rate,
-  #remove the id field
-  if(id_field ==  TRUE){
-    dat  <- dat[,2:ncol(dat)]
-    c_name <- colnames(traj)[1]
+#back up data
+b_dat <- dat
+#remove the id field
+if(id_field ==  TRUE){
+  dat  <- dat[,2:ncol(dat)]
+  c_name <- colnames(traj)[1]
+}
+
+#matrix to track the outlier incidents [TRUE or FALSE]
+outlier_mat <- matrix(FALSE, nrow(dat), ncol(dat))
+
+#------------------------------------
+#if method: "quantile"
+if(method==1){
+  #check if the value is in-between [0,1]
+  if(threshold < 0 | threshold > 1){
+    stop("*--Terminated!!!--*, The 'threshold' value should be between 0 and 1.")
   }
+#calculate the cut-off value based on the 'threshold'
+thres_ <- as.vector(round(quantile(as.vector(unlist(as.data.frame(dat))),
+                                       threshold), digits=5))
+id_ <- which(dat > thres_)
+  #update the outlier tracker
+  outlier_mat[id_] <- "TRUE"
+}
 
-  #matrix to track the outlier incidents [TRUE or FALSE]
-  outlier_mat <- matrix(FALSE, nrow(dat), ncol(dat))
+# method: "manual"
+#------------------------------------
+if(method==2){
+  id_ <- which(dat > threshold)
+  outlier_mat[id_] <- "TRUE"
+}
 
-  #------------------------------------
-  #if method: "quantile"
-  if(method==1){
-    #check if the value is in-between [0,1]
-    if(threshold < 0 | threshold > 1){
-      stop("*--Terminated!!!--*, The 'threshold' value should be between 0 and 1.")
-    }
-    #calculate the cut-off value based on the 'threshold'
-    thres_ <- as.vector(round(quantile(as.vector(unlist(as.data.frame(dat))), threshold), digits=5))
-    id_ <- which(dat > thres_)
-    #update the outlier tracker
-    outlier_mat[id_] <- "TRUE"
+#check if an entire row is ouliers...then terminate.
+c_out <- NULL
+for(n_ in 1:ncol(outlier_mat)){ #n_<-1
+  c_out <- rbind(c_out, cbind(n_, length(which(outlier_mat[,n_]==TRUE))))
+}
+
+if(replace_with == 1){
+  wc_out <- which(c_out[,2]==nrow(dat))
+  if(length(wc_out)!=0){
+    stop(paste("*--Function terminated!!!--* All observations on Column(s)",
+               wc_out, "are outliers!! Please, set a higher 'threshold' value!!", sep=" "))
   }
+}
 
-  # method: "manual"
-  #------------------------------------
-  if(method==2){  #threshold=17
-    id_ <- which(dat > threshold)
-    outlier_mat[id_] <- "TRUE"
+#check if an entire row is ouliers...then terminate.
+r_out <- NULL
+for(m_ in 1:nrow(outlier_mat)){ #m_<-1
[…]
