+++ R/dataImputation.R
@@ -25,43 +25,43 @@
 #' @return A data.frame with missing values (\code{NA}, \code{Inf}, \code{null}) imputed according to the a specified technique.
 #' @export
 
-dataImputation <- function(traj, id_field = FALSE, method = 2, replace_with = 1, fill_zeros = FALSE){
+dataImputation <- function(traj, id_field = FALSE,
+                           method = 2, replace_with = 1,
+                           fill_zeros = FALSE){
 
-  dat <- as.data.frame(traj)
+dat <- as.data.frame(traj)
 
-  coln_ <- colnames(dat)
+coln_ <- colnames(dat)
 
-  #check if there is unique(id) field
-  if(id_field==TRUE){
-    backup_dat <- dat
-    dat <- dat[,2:ncol(dat)]
-  }
+#check if there is unique(id) field
+if(id_field==TRUE){
+  backup_dat <- dat
+  dat <- dat[,2:ncol(dat)]
+}
 
-  #matrix to backup the 'null' information store via the outlier information [TRUE or FALSE]
-  null_mat <- matrix(FALSE, nrow(dat), ncol(dat))
+#matrix to backup the 'null' information store via the outlier information [TRUE or FALSE]
+null_mat <- matrix(FALSE, nrow(dat), ncol(dat))
 
-  for(a in 1:ncol(dat)){ #a<-1
-    sub_ <- suppressWarnings(as.numeric(as.character(dat[a,])))
-    i_ <- which(is.na(sub_))
-    j_ <- which(is.infinite(sub_))
-    k_ <- which(is.null(sub_))
-    i_m <- unique(c(i_, j_, k_))[order(unique(c(i_, j_, k_)))]
+for(a in seq_len(ncol(dat))){
+  sub_ <- suppressWarnings(as.numeric(as.character(dat[a,])))
+  i_ <- which(is.na(sub_))
+  j_ <- which(is.infinite(sub_))
+  k_ <- which(is.null(sub_))
+  i_m <- unique(c(i_, j_, k_))[order(unique(c(i_, j_, k_)))]
+  null_mat[a,i_m] <- TRUE
+}
 
-    null_mat[a,i_m] <- TRUE
-  }
-
-  # arithmetic method
-  if(method==1){
-    #Use column values in order to determine the new value.
-    if(replace_with==1 | replace_with==2 | replace_with==3){
-      fill_count <- 0
-      for(g in 1:ncol(null_mat)){ #g<-2
-        i_m <- which(null_mat[,g]==TRUE)
-        subset_ <- dat[-i_m,g]
-
-        if(length(i_m)!=0){fill_count <- fill_count + length(i_m)}
-
-        if(length(i_m)!=0){
+#arithmetic method
+if(method==1){
+  #Use column values in order to determine the new value.
+  if(replace_with==1 | replace_with==2 | replace_with==3){
+    fill_count <- 0
+    for(g in seq_len(ncol(null_mat))){ #g<-2
+      i_m <- which(null_mat[,g]==TRUE)
+      subset_ <- dat[-i_m,g]
+      if(length(i_m)!=0){fill_count <-
+        fill_count + length(i_m)}
+      if(length(i_m)!=0){
           if(replace_with==1){
             ave_ <- mean(subset_)
           }
@@ -72,86 +72,82 @@ dataImputation <- function(traj, id_field = FALSE, method = 2, replace_with = 1,
             ave_ <- max(subset_)
           }
           dat[i_m,g] <- ave_
-        }
       }
     }
+  }
 
-    #Use row values in order to determine the new value.
-    if(replace_with==4 | replace_with==5 | replace_with==6){
-      fill_count <- 0
[…]
