+++ R/alphaLabel.R
@@ -13,12 +13,13 @@
 alphaLabel <- function(x){
 
   combind_A <- LETTERS
-  combind <-  combn(LETTERS, m=2, sep="")# combind[1:2,]
+  combind <-  combn(LETTERS, m=2, sep="")
   list_Letters <- NULL
   for(cc in 1:ncol(combind)){#cc=1
-    list_Letters <-c(list_Letters,  paste(combind[1,cc],   combind[2,cc], sep=""))
+    list_Letters <-c(list_Letters,  paste(combind[1,cc],
+                                          combind[2,cc], sep=""))
   }
-  list_Letters <- c(combind_A, list_Letters) #combine
+  list_Letters <- c(combind_A, list_Letters)
   if(length(unique(x))<=350){
     clust_num <- list_Letters[x]
     return(clust_num)
