### 2020-02-07 58457d0
update alphaLabel.R
