### 2020-02-07 1b4f5b1
update clustering function
