+++ NAMESPACE
@@ -7,6 +7,7 @@ export(elbowPoint)
 export(outlierDetect)
 export(props)
 export(rates)
+export(removeRowsN)
 export(statPrint)
 export(wSpaces)
 import(kml, grDevices, reshape2, Hmisc, stats, utils, ggplot2, longitudinalData)
