### 2020-02-25 #2019 @MAnalytics
@karthik Thank you very much. I hope to hear from you soon.
