### 2020-06-22 #2019 @MAnalytics
Dear @karthik
Regarding the reminders you sent, did they get back to you? If no, when would you commence the review as suggested.
Thank you as I anticipate your kind response.
