### 2020-06-23 #2379 @whedon
[REVIEW]: Akmedoids R package for generating directionally-homogeneous clusters of longitudinal data sets
**Submitting author:** @MAnalytics (<a href="http://orcid.org/0000-0002-9006-4934">Monsuru Adepeju</a>)
**Repository:** <a href="https://github.com/MAnalytics/akmedoids" target ="_blank">https://github.com/MAnalytics/akmedoids</a>
**Version:** v2.0.0
