### 2020-07-05 #2379 @MAnalytics
> Just touching base here, I'll review this article this week!

Thank you very much. I appreciate it.
