### 2020-07-29 #2379 @karthik
### Initial review of Akmedoids R package for generating directionally-homogeneous clusters of longitudinal data sets

Here are some general heuristic checks. Particularly problematic issues are using `1:nrow` and importing full packages.

```
── Good Practice akmedoids

It is good practice to

    R/akmedoids.clust.R:31:NA
    R/akmedoids.clust.R:35:NA
    R/akmedoids.clust.R:36:NA
    R/akmedoids.clust.R:40:NA
    R/akmedoids.clust.R:43:NA
    ... and 658 more lines

  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easier
    to read your code for them if you use '<-'.

    R/akmedoids.clust.R:77:8
    R/akmedoids.clust.R:83:6
    R/akmedoids.clust.R:245:14
    R/akmedoids.clust.R:272:23
    R/akmedoids.clust.R:295:23
    ... and 22 more lines

    R/akmedoids.clust.R:3:1
    R/akmedoids.clust.R:4:1
    R/akmedoids.clust.R:5:1
    R/akmedoids.clust.R:6:1
    R/akmedoids.clust.R:7:1
    ... and 106 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
    result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

    R/akmedoids.clust.R:77:10
    R/akmedoids.clust.R:81:22
    R/akmedoids.clust.R:89:34
    R/dataImputation.R:109:8
    R/outlierDetect.R:67:11
    ... and 16 more lines

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
  ✖ avoid 'T' and 'F', as they are just variables which are
    set to the logicals 'TRUE' and 'FALSE' by default, but are not
    reserved words and hence can be overwritten by the user.  Hence,
    one should always use 'TRUE' and 'FALSE' for the logicals.

    R/statPrint.R:NA:NA
```

- Please add a language field to your DESCRIPTION file

- Please check your package for spelling mistakes. Not all of these are mistakes but I see a few.

 <summary>Potential Spelling mistakes</summary>
<pre><code>
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
  WORD              FOUND IN
’Whitespaces’ wSpaces.Rd:19
                  akmedoids-vignette.Rmd:365
agentlans         elbowPoint.Rd:21
al                akmedoids.clust.Rd:39
                  statPrint.Rd:27,57
                  description:2
                  README.md:7
                  akmedoids-vignette.Rmd:358,428,526,567,605
alphaLabel        alphaLabel.Rd:19
analysing         README.md:3
Appl              akmedoids.clust.Rd:40
behavioural       akmedoids-vignette.Rmd:126
Calinski          akmedoids.clust.Rd:18,24,41
                  description:6
Caliński         README.md:104
categorise        statPrint.Rd:27
                  akmedoids-vignette.Rmd:536
centroids         akmedoids.clust.Rd:27
charater          statPrint.Rd:19
Commun            akmedoids.clust.Rd:41
complementarity   akmedoids-vignette.Rmd:504
Comput            akmedoids.clust.Rd:40
corresp           dataImputation.Rd:17,18,19,20,26
customised        statPrint.Rd:43
                  akmedoids-vignette.Rmd:578
denomin           rates.Rd:14
doi               description:2,7,10
et                akmedoids.clust.Rd:39
                  statPrint.Rd:27,57
                  description:2
                  README.md:7
                  akmedoids-vignette.Rmd:358,428,526,567,605
GBTM              akmedoids-vignette.Rmd:126
Genolini          akmedoids.clust.Rd:39
                  description:2
ggplot            description:9
github            elbowPoint.Rd:21
Harabasz          akmedoids.clust.Rd:24,41
                  README.md:104
Harabatz          akmedoids.clust.Rd:18
                  description:6
http              akmedoids.clust.Rd:39
imput             akmedoids-vignette.Rmd:194
initialisation    akmedoids.clust.Rd:14
                  akmedoids-vignette.Rmd:126
jss               description:2
jstatsoft         akmedoids.clust.Rd:39
judgement         akmedoids-vignette.Rmd:504
kmeans            README.md:7
                  akmedoids-vignette.Rmd:126
Kmedoids          title:1
kml               akmedoids.clust.Rd:39
knitr             akmedoids-vignette.Rmd:428
maximisation      akmedoids-vignette.Rmd:126
md                NEWS.md:2
medoid            statPrint.Rd:25,27
                  akmedoids-vignette.Rmd:526,534,536,567
minimised         akmedoids-vignette.Rmd:126
minimises         akmedoids-vignette.Rmd:126
modelling         akmedoids-vignette.Rmd:126
normalise         akmedoids-vignette.Rmd:262
optimality        akmedoids.clust.Rd:18
param             akmedoids.clust.Rd:21
png               akmedoids-vignette.Rmd:428
propotion         props.Rd:14
quartile          statPrint.Rd:27
recognise         akmedoids-vignette.Rmd:194
recognises        akmedoids.clust.Rd:12
                  dataImputation.Rd:12
                  outlierDetect.Rd:13
                  props.Rd:12
                  removeRowsN.Rd:12
                  statPrint.Rd:23
                  akmedoids-vignette.Rmd:163
removeRowNA       NEWS.md:13
Rousseeuw         akmedoids.clust.Rd:40
                  description:5
                  README.md:102
sep               akmedoids-vignette.Rmd:428
showplots         NEWS.md:12
spacebar          wSpaces.Rd:19
                  akmedoids-vignette.Rmd:365
statPrints        NEWS.md:12
traj              rates.Rd:12,14
vartheta          akmedoids-vignette.Rmd:526
ve                akmedoids-vignette.Rmd:567
victimisation     akmedoids-vignette.Rmd:126
visualisation     statPrint.Rd:43
                  description:8
                  akmedoids-vignette.Rmd:578
visualise         akmedoids-vignette.Rmd:362
Wickham           description:9
www               akmedoids.clust.Rd:39
</code></pre>

Your tests are on a folder called `test` NOT `tests` but uses `testthat` infrastructure. How are you running the tests? Also your tests are not passing and barely cover any of the code. Please fix your tests and coverage before we can proceed.
