+++ NEWS.md
@@ -13,3 +13,5 @@ output: html_document
 2. Added a new function 'removeRowNA.R'
 
 
+
+
