+++ DESCRIPTION
@@ -33,7 +33,7 @@ Imports:
     longitudinalData, 
     stats, 
     signal
-RoxygenNote: 7.0.1
+RoxygenNote: 7.1.0
 Suggests:
     knitr,
     rmarkdown,
@@ -43,5 +43,3 @@ Suggests:
     dplyr,
     digest
 VignetteBuilder: knitr
-
-
