+++ vignettes/akmedoids-vignette.Rmd
@@ -1,609 +1,0 @@
----
-title: "A guide to measuring long-term inequality in the exposure to crime at micro-area levels using `'Akmedoids'` package"
-
-author: |
-  | `Authors:`
-  | `Adepeju, M., Langton, S., and Bannister, J.`
-  | `Big Data Centre, Manchester Metropolitan University, Manchester, M15 6BH`
-  
-date: |
-  | `Date:`
-  | ``r Sys.Date()``
-
-output:
-  rmarkdown::html_vignette
-  
-#dev: png
-#output:
-  #word_document: default
-  #always_allow_html: yes
-#  pdf_document: default
-always_allow_html: yes
-#fig_caption: yes
-bibliography: references.bib
-
-abstract: The `'akmedoids'` package advances a set of R-functions for longitudinal clustering of long-term trajectories and determines the optimal solution based on the `Caliński-Harabatz` criterion (Caliński and Harabasz 1974). The package also includes a set of functions for addressing common data issues, such as missing entries and outliers, prior to conducting advance longitudinal data analysis. One of the key objectives of this package is to facilitate easy replication of a recent paper which examined small area inequality in the crime drop (see Adepeju et al. 2019). This document is created to provide a guide towards accomplishing this objective. Many of the functions provided in the `akmedoids` package may be applied to longitudinal data in general.
-  
-vignette: >
-  %\VignetteIndexEntry{A guide to measuring long-term inequality in the exposure to crime at micro-area levels using 'Akmedoids' package}
-  %\VignetteEngine{knitr::rmarkdown}
-  \usepackage[utf8]{inputenc}
----
-
-<style type="text/css">
-
-h1.title {
-  font-size: 26px;
-  line-height: 130%;
-  color: Black;
-  text-align: center;
-}
-
-h2.subtitle {
-  font-size: 13px;
-  line-height: 120%;
-  color: Black;
-  text-align: center;
-}
-
-h4.author { /* Header 4 - and the author and data headers use this too  */
-  font-size: 17px;
-  font-family: "Arial";
-  color: Black;
-  text-align: center;
-}
-h4.date { /* Header 4 - and the author and data headers use this too  */
-  font-size: 17px;
-  font-family: "Arial", Times, serif;
-  color: Black;
-  text-align: center;
-}
-
-h4.abstract { /* Header 4 - and the author and data headers use this too  */
-  font-size: 10px;
-  font-family: "Arial", Times, serif;
-  color: black;
-  text-align: center;
-}
-
-h4.institute{ /* Header 4 - and the author and data headers use this too  */
-  font-size: 10px;
-  font-family: "Arial", Times, serif;
-  color: black;
-  text-align: center;
-}
-
-body, td {
-   font-size: 14px;
-}
-code.r{
-  font-size: 13px;
-}
-pre {
-  font-size: 13px
-}
-h1 { /* Header 1 */
-  font-size: 16px;
-  color: DarkBlue;
-}
-h2 { /* Header 2 */
-    font-size: 16px;
-  color: DarkBlue;
-}
-h3 { /* Header 3 */
-  font-size: 15px;
-  font-family: "Times New Roman", Times, serif;
-  color: DarkBlue;
-
-</style>
-
-```{r setup, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-```{r functions, include=FALSE}
-# A function for captioning and referencing images
-fig <- local({
-    i <- 0
-    ref <- list()
-    list(
-        cap=function(refName, text) {
-            i <<- i + 1
-            ref[[refName]] <<- i
-            paste("Figure ", i, ": ", text, sep="")
-        },
-        ref=function(refName) {
-            ref[[refName]]
-        })
-})
-```
-
-# Introduction
-
[…]
