### 2020-08-02 adbe2df
clustering function
