### 2020-08-04 44d2585
clustering function
