### 2020-08-04 265849c
print stats
