### 2020-08-05 830f03a
clustering function
