+++ man/clustr.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/clustr.R
+\docType{data}
+\name{clustr}
+\alias{clustr}
+\title{Sample clusters}
+\format{
+A vector
+}
+\usage{
+clustr
+}
+\description{
+A vector of cluster labels
+}
+\keyword{datasets}
