+++ man/dataImputation.Rd
@@ -14,16 +14,16 @@ dataImputation(traj, id_field = FALSE, method = 2, replace_with = 1, fill_zeros 
 \item{method}{[an integer] indicating a method for calculating the missing values. Options are: \code{'1'}: \code{arithmetic} method, and \code{'2'}: \code{regression} method. The default is \code{'1'}: \code{arithmetic} method}
 
 \item{replace_with}{[an integer from 1 to 6] indicating the technique, based on a specified \code{method}, for calculating the missing entries.
-\code{'1'}: \code{arithmetic} method, \code{replace_with} options are: \code{'1'}: Mean value of the corresp column;
-\code{'2'}: Minimum value of corresp column; \code{'3'}: Maximum value of corresp column;
-\code{'4'}: Mean value of corresp row; \code{'5'}: Minimum value of corresp row,
-or \code{'6'}: Maximum value of corresp row. For \code{'2'}: regression method:
+\code{'1'}: \code{arithmetic} method, \code{replace_with} options are: \code{'1'}: Mean value of the corresponding column;
+\code{'2'}: Minimum value of corresponding column; \code{'3'}: Maximum value of corresponding column;
+\code{'4'}: Mean value of corresponding row; \code{'5'}: Minimum value of corresponding row,
+or \code{'6'}: Maximum value of corresponding row. For \code{'2'}: regression method:
 the available option for the \code{replace_with} is: \code{'1'}: \code{linear}.
 The regression method fits a linear regression line to a trajectory with missing entry(s)
 and estimates the missing data values from the regression line.
 Note: only the missing data points derive their new values from the regression line
 while the rest of the data points retain their original values. The function terminates if there are
-trajectories with only one observation. The default is \code{'1'}: Mean value of the corresp column}
+trajectories with only one observation. The default is \code{'1'}: Mean value of the corresponding column}
 
 \item{fill_zeros}{[TRUE or FALSE] whether to consider zeros \code{0} as missing values when \code{2: regression} method is used. The default is \code{FALSE}.}
 }
