+++ man/statPrint.Rd
@@ -16,45 +16,99 @@ statPrint(
 )
 }
 \arguments{
-\item{clustr}{[vector (charater)] A vector of cluster membership (labels). For instance, the result extracted from the \code{\link{akmedoids.clust}} function.}
+\item{clustr}{[vector (character)] A vector of cluster
+membership (labels). For instance, the result extracted
+from the \code{\link{akmedoids.clust}} function.}
 
-\item{traj}{[matrix (numeric)]: corresponding longitudinal data used to generate \code{clustr} (with rows corresponding to each label of \code{clustr}). For example, the first label of \code{clustr} is the group label of the first row of \code{traj} matrix, and so on.}
+\item{traj}{[matrix (numeric)]: corresponding longitudinal
+data used to generate \code{clustr} (with rows
+corresponding to each label of \code{clustr}).
+For example, the first label of \code{clustr}
+is the group label of the first row of
+\code{traj} matrix, and so on.}
 
-\item{id_field}{[numeric or character] Whether the first column of the \code{traj} is a unique (\code{id}) field. Default: \code{FALSE}. If \code{TRUE} the function recognises the second column as the first time step.}
+\item{id_field}{[numeric or character] Whether the first
+column of the \code{traj} is a unique (\code{id}) field.
+Default: \code{FALSE}. If \code{TRUE} the function
+recognises the second column as the first time step.}
 
-\item{reference}{[numeric] Specifying the reference line from which the direction of each group is measured. Options are: \code{1}: slope of mean trajectory, \code{2}: slope of medoid trajectory, \code{3}: slope of a horizontal line (i.e. slope = 0). Default: \code{1}.}
+\item{reference}{[numeric] Specifying the reference line from
+which the direction of each group is measured. Options are:
+\code{1}: slope of mean trajectory, \code{2}: slope of medoid
+trajectory, \code{3}: slope of a horizontal line
+(i.e. slope = 0). Default: \code{1}.}
 
-\item{N.quant}{[numeric] Number of equal intervals (quantiles) to create between the reference line \code{(R)} and the medoids \code{(M)} of the most-diverging groups of both sides of \code{(R)}. Default is \code{4} - meaning quartile subdivisions on each side of \code{(R)}. In this scenario, the function returns the quartile in which the medoid of each group falls. This result can be used to further categorise the groups into 'classes'. For example, groups that fall within the \code{1st} quartile may be classified as 'Stable' groups (Adepeju et al. 2019).}
+\item{N.quant}{[numeric] Number of equal intervals (quantiles)
+to create between the reference line \code{(R)} and the medoids
+\code{(M)} of the most-diverging groups of both sides of
+\code{(R)}. Default is \code{4} - meaning quartile subdivisions
+on each side of \code{(R)}. In this scenario, the function
+returns the quartile in which the medoid of each group falls.
+This result can be used to further categorize the groups into
+'classes'. For example, groups that fall within the \code{1st}
+quartile may be classified as 'Stable' groups (Adepeju et al. 2019).}
 
-\item{showplots}{[TRUE or FALSE] To display cluster plots. Defaults \code{TRUE}}
+\item{showplots}{[TRUE or FALSE] To display cluster plots.
+Defaults \code{TRUE}}
 
-\item{type}{[character] plot type. Available options are: \code{"lines"} and \code{"stacked"}.}
+\item{type}{[character] plot type. Available options are:
+\code{"lines"} and \code{"stacked"}.}
 
[…]
