### 2020-08-10 b6d70df
Update traj.Rd
