### 2020-08-10 8c1b206
Update wSpaces.Rd
