### 2020-08-10 589beeb
Update akmedoids.clust.R
