### 2020-08-10 b42a4fd
Update statPrint.R
