+++ R/statPrint.R
@@ -1,116 +1,188 @@
 
 #' @title Descriptive (Change) statistics and plots
-#' @description This function perform two tasks: (i) it generate the descriptive and change statistics of groups, particularly suited for the outputs form the \code{\link{akmedoids.clust}} function, and (ii) generates the plots of the groups (performances).
-#' @param clustr [vector (character)] A vector of cluster membership (labels). For instance, the result extracted from the \code{\link{akmedoids.clust}} function.
-#' @param traj [matrix (numeric)]: corresponding longitudinal data used to generate \code{clustr} (with rows corresponding to each label of \code{clustr}). For example, the first label of \code{clustr} is the group label of the first row of \code{traj} matrix, and so on.
-#' @param id_field [numeric or character] Whether the first column of the \code{traj} is a unique (\code{id}) field. Default: \code{FALSE}. If \code{TRUE} the function recognises the second column as the first time step.
-#' @param type [character] plot type. Available options are: \code{"lines"} and \code{"stacked"}.
-#' @param y.scaling [character] works only if \code{type="lines"}. \code{y.scaling} set the vertical scales of the cluster panels. Options are: \code{"fixed"}: uses uniform scale for all panels, \code{"free"}: uses variable scales for panels.
-#' @param reference [numeric] Specifying the reference line from which the direction of each group is measured. Options are: \code{1}: slope of mean trajectory, \code{2}: slope of medoid trajectory, \code{3}: slope of a horizontal line (i.e. slope = 0). Default: \code{1}.
-#' @param N.quant [numeric] Number of equal intervals (quantiles) to create between the reference line \code{(R)} and the medoids \code{(M)} of the most-diverging groups of both sides of \code{(R)}. Default is \code{4} - meaning quartile subdivisions on each side of \code{(R)}. In this scenario, the function returns the quartile in which the medoid of each group falls. This result can be used to further categorize the groups into 'classes'. For example, groups that fall within the \code{1st} quartile may be classified as 'Stable' groups (Adepeju et al. 2019).
-#' @param showplots [TRUE or FALSE] To display cluster plots. Defaults \code{TRUE}
+#' @description This function perform two tasks:
+#' (i) it generate the descriptive and change statistics
+#' of groups, particularly suited for the outputs form
+#' the \code{\link{akmedoids.clust}} function, and
+#' (ii) generates the plots of the groups (performances).
+#' @param clustr [vector (character)] A vector of cluster
+#' membership (labels). For instance, the result extracted
+#' from the \code{\link{akmedoids.clust}} function.
+#' @param traj [matrix (numeric)]: corresponding longitudinal
+#' data used to generate \code{clustr} (with rows
+#' corresponding to each label of \code{clustr}).
+#' For example, the first label of \code{clustr}
+#' is the group label of the first row of
+#' \code{traj} matrix, and so on.
+#' @param id_field [numeric or character] Whether the first
+#' column of the \code{traj} is a unique (\code{id}) field.
+#' Default: \code{FALSE}. If \code{TRUE} the function
+#' recognises the second column as the first time step.
[…]
