### 2020-08-10 63b3333
Update traj.R
