+++ R/traj.R
@@ -1,5 +1,6 @@
 #' @title Longitudinal dataset
-#' @description Simulated longitudinal datasets with missing values (\code{NA}, \code{Inf}, \code{null})
+#' @description Simulated longitudinal datasets with
+#' missing values (\code{NA}, \code{Inf}, \code{null})
 #' @name traj
 #' @docType data
 #' @usage traj
