### 2020-08-10 7feec0d
Update wSpaces.R
