### 2020-08-10 2136fe7
Delete testthat.R
