+++ test/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(akmedoids)
-test_check("akmedoids")
-
