### 2020-08-10 370c08d
Delete test-clustering.R
