### 2020-08-10 8ebee19
Delete test-manipulate.R
