### 2020-08-14 97994e6
update row removal function
