+++ man/removeRowsN.Rd
@@ -2,25 +2,37 @@
 % Please edit documentation in R/removeRowsN.R
 \name{removeRowsN}
 \alias{removeRowsN}
-\title{Removes rows that contain 'NA' and 'Inf' entries}
+\title{Removes rows that contain 'NA' and/or 'Inf' entries}
 \usage{
-removeRowsN(traj, id_field=TRUE)
+removeRowsN(traj, id_field=TRUE, remove=1)
 }
 \arguments{
-\item{traj}{[data.frame (numeric)]: longitudinal data. Each row represents an individual trajectory (of observations). The columns show the observations at consecutive time points.}
+\item{traj}{[data.frame (numeric)]: longitudinal data.
+Each row represents an individual trajectory (of observations).
+The columns show the observations at consecutive time points.}
 
-\item{id_field}{[numeric or character] Whether the first column of the \code{traj} is a unique (\code{id}) field. Default: \code{FALSE}. If \code{TRUE} the function recognises the second column as the first time step.}
+\item{id_field}{[numeric or character] Whether the first column
+of the \code{traj} is a unique (\code{id}) field.
+Default: \code{FALSE}. If \code{TRUE} the function recognises
+the second column as the first time step.}
+
+\item{remove}{[integer] Type of missing entries to remove.
+\code{1} for 'NA', \code{2} for 'Inf', and \code{3} for both.
+Default:\code{1}.}
 }
 \value{
 A matrix with complete observations
 }
 \description{
-This function removes any rows in which an 'NA' or an 'Inf' entry is found.
+This function removes any rows in which an 'NA'
+or an 'Inf' entry is found.
 }
 \details{
-Given a matrix (or a dataframe) containing an 'NA' or an 'Inf' entry, the function returns only rows with complete observations.
+Given a matrix (or a dataframe) containing an 'NA' or
+an 'Inf' entry, the function returns only rows with
+complete observations.
 }
 \examples{
 traj <- traj
-removeRowsN(traj, id_field=TRUE)
+removeRowsN(traj, id_field=TRUE, remove=3)
 }
