### 2020-08-15 9026731
update wSpaces.Rd
