+++ man/wSpaces.Rd
@@ -4,12 +4,16 @@
 \alias{wSpaces}
 \title{Whitespaces removal}
 \usage{
-wSpaces(traj)
+wSpaces(traj, remove="Both")
 }
 \arguments{
 \item{traj}{[matrix (numeric)]: longitudinal data. Each row
 represents an individual trajectory (of observations).
 The columns show the observations at consecutive time points.}
+
+\item{remove}{[character]: Type of whitespace to remove.
+That is, "Left" (leading), (2) "Right" (trailing), or "Both"
+(both leading and trailing whitespaces). Default: "Both".}
 }
 \value{
 A matrix with all whitespaces (if any) removed.
@@ -20,17 +24,17 @@ trailing whitespaces in data
 }
 \details{
 Given a matrix suspected to contain whitespaces,
-this function removes all the whitespaces and returns a
+this function removes the type of the whitespaces specified and returns a
 cleaned data. ’Whitespaces’ are white characters often
 introduced into data during data entry, for instance by
 wrongly pressing the spacebar. For example, neither " A"
-nor "A " equates "A" because of the whitespaces that
+nor "A " is the same as "A" because of the whitespaces that
 exist in them. They can also result from systematic
 errors in data recording devices.
 }
 \examples{
 traj <- traj
-wSpaces(traj)
+wSpaces(traj, remove="Both")
 }
 \references{
 \url{https://en.wikipedia.org/wiki/Whitespace_character}
