### 2020-08-15 41eeecd
Update removeRowsN.R
