+++ R/wSpaces.R
@@ -8,20 +8,20 @@
 #' @param remove [character]: Type of whitespace to remove.
 #' That is, "Left" (leading), (2) "Right" (trailing), or "Both"
 #' (both leading and trailing whitespaces). Default: "Both".
-#' @usage wSpaces(traj)
+#' @usage wSpaces(traj, remove="Both")
 #' @details Given a matrix suspected to contain whitespaces,
 #' this function removes the type of the whitespaces specified and returns a
 #' cleaned data. ’Whitespaces’ are white characters often
 #' introduced into data during data entry, for instance by
 #' wrongly pressing the spacebar. For example, neither " A"
-#' nor "A " equates "A" because of the whitespaces that
+#' nor "A " is the same as "A" because of the whitespaces that
 #' exist in them. They can also result from systematic
 #' errors in data recording devices.
 #' @return A matrix with all whitespaces (if any) removed.
 #' @references \url{https://en.wikipedia.org/wiki/Whitespace_character}
 #' @examples
 #' traj <- traj
-#' wSpaces(traj)
+#' wSpaces(traj, remove="Both")
 #' @export
 
 wSpaces <- function(traj, remove="Both"){
@@ -32,28 +32,30 @@ wSpaces <- function(traj, remove="Both"){
 
   coln_ <- colnames(dat_Cleaned)
 
-  count_trailing <- 0 #keep the count of whitespace removed.
-  count_leading <- 0 #keep the count of whitespace removed.
+  count_trailing <- 0
+  count_leading <- 0
+
+  solution <- list()
 
   for(q in seq_len(ncol(dat_Cleaned))){
 
     #dat_Cleaned=test.data1
 
-    if(remove="Right"){
+    if(remove=="Right"){
       #remove trailing whitespaces
       vec_Name1 <- trimws(as.vector(dat_Cleaned[,q]), which="right")
       #count the number of trailing whitespaces
       count_trailing <- count_trailing + length(which(!vec_Name1%in%dat_Cleaned[,q]))
     }
 
-    if(remove="Left"){
+    if(remove=="Left"){
       #remove leading whitespaces
       vec_Name1 <- trimws(as.vector(dat_Cleaned[,q]), which="left")
       #count the number of leading whitespaces
       count_leading <- count_leading + length(which(!vec_Name1%in%dat_Cleaned[,q]))
     }
 
-    if(remove="Both"){
+    if(remove=="Both"){
       #remove trailing whitespaces
       vec_Name1 <- trimws(as.vector(dat_Cleaned[,q]), which="right")
       #count the number of trailing whitespaces
@@ -74,21 +76,24 @@ wSpaces <- function(traj, remove="Both"){
 
   flush.console()
 
-  if(remove="Right"){
+  if(remove=="Right"){
     print(paste(count_trailing, "trailing whitespaces found/removed!"))
+    Removed <- count_trailing
   }
 
-  if(remove="Left"){
+  if(remove=="Left"){
     print(paste(count_leading, "leading whitespaces found/removed!"))
+    Removed <- count_leading
   }
 
-  if(remove="Both"){
+  if(remove=="Both"){
     print(paste((count_leading + count_trailing), "leading whitespaces found/removed!"))
+    Removed <- count_leading + count_trailing
   }
 
-  dat_Cleaned = data.frame(dat_Cleaned)
+  solution <- list(NumberOfWhiteSpacesRemoved=Removed, CleanData=dat_Cleaned)
 
-  return(dat_Cleaned)
+  return(solution)
 
 }
 
