### 2020-08-16 cf74257
update rates.R
