+++ R/rates.R
@@ -18,8 +18,7 @@
 #' \code{traj/denomin} is expressed. Default is \code{100}.
 #' @usage rates(traj, denomin, id_field, multiplier)
 #' @examples
-#' traj2 <- traj
-#' traj2 <- dataImputation(traj2, id_field = TRUE, method = 2,
+#' traj2 <- dataImputation(traj, id_field = TRUE, method = 2,
 #' replace_with = 1, fill_zeros = FALSE)
 #' pop <- population #read denominator data
 #' pop2 <- as.data.frame(matrix(0, nrow(population), ncol(traj)))
@@ -44,64 +43,93 @@
 
 rates <- function(traj, denomin, id_field=TRUE, multiplier = 100){
 
-dat1 <- traj
-dat2 <- denomin
+  solution <- list()
+  dat1 <- traj
+  dat2 <- denomin
 
-#compare the number of columns
-if(ncol(traj)!=ncol(denomin)){
-  stop("*---Number of columns must be the same---*")
-}
-
-#compare the number of columns
-if(id_field==FALSE){
-  stop("*---unique field must be set as 'TRUE'!---*")
-}
-
-#check uniqueness of the fields
-if(id_field==TRUE){
-  n_CL <- colnames(dat1)
-  col_names1 <- as.vector(as.character(dat1[,1]))
-  col_names2 <- as.vector(as.character(dat2[,1]))
-
-  #dat <- dat[,2:ncol(dat)]
-  #check if the "id_field" is a unique field
-  if(!length(col_names1)==length(unique(col_names1))){
-    stop("(: The 'id_field' of the 'traj' object is not a unique field. Function terminated!!! :)")
+  #examine the number of columns
+  if(ncol(traj) < 3 | ncol(denomin) < 3){
+    stop("*---Number of columns of dataset must be greater than 3---*")
   }
-  if(!length(col_names2)==length(unique(col_names2))){
+
+  #compare the number of columns
+  if(ncol(traj)!=ncol(denomin)){
+    stop("*---Number of columns must be the same---*")
+  }
+
+  #compare the number of columns
+  if(id_field==FALSE){
+    stop("*---unique field must be set as 'TRUE'!---*")
+  }
+
+  #check uniqueness of the fields
+  if(id_field==TRUE){
+    n_CL <- colnames(dat1)
+    id_names1 <- as.vector(as.character(dat1[,1]))
+    id_names2 <- as.vector(as.character(dat2[,1]))
+
+    if(!length(id_names1)==length(unique(id_names1))){
+      stop("(: The 'id_field' of the 'traj' object is not a unique field. Function terminated!!! :)")
+    }
+    if(!length(id_names2)==length(unique(id_names2))){
       stop("(: The 'id_field' of the 'denominator' object is not a unique field. Function terminated!!! :)")
+    }
   }
-}
 
+  #----------------------------------------------------------
+  #trajectory data
+  #----------------------------------------------------------------------
   data1 <- apply(dat1[,2:ncol(dat1)], 2, as.numeric)#head(data1)
-  data1 <- cbind(1:nrow(data1), data1)
-  colnames(data1) <- c("ID", 1:(ncol(data1)-1))
+  data1 <- cbind(seq_len(nrow(data1)), data1)
+  colnames(data1) <- c("ID", seq_len((ncol(data1)-1)))
+
   #----------------------------------------------------------
   #denominator data
   #----------------------------------------------------------------------
   data2 <- apply(dat2[,2:ncol(dat2)], 2, as.numeric)#head(data2)
[…]
