### 2020-08-16 e6d0871
update test_rates.R
