### 2020-08-18 1d3fefb
update test_rates.R
