### 2020-08-18 dfc8e39
test_props.R updates
