### 2020-08-18 1c2236c
updates test_outlierDetect.R
