### 2020-08-18 1c547cd
update outlierDetect
