### 2020-08-18 43471e4
updates elbowPoint
