+++ R/alphaLabel.R
@@ -1,10 +1,14 @@
 #' @title Numerics ids to alphabetical ids
-#' @description Function to transform a list of numeric ids to alphabetic ids
+#' @description Function to transform a list of numeric ids
+#' to alphabetic ids
 #' @param x A vector of numeric ids
 #' @usage alphaLabel(x)
-#' @details Given a vector of numeric cluster ids, `alphaLabel` converts each id to its corresponding alphabets. It combines alphabets for ids greater than 26.
+#' @details Given a vector of numeric cluster ids,
+#' `alphaLabel` converts each id to its corresponding alphabets.
+#' It combines alphabets for ids greater than 26.
 #' @return A vector of alphabetical ids.
 #' @examples
+#' set.seed(10)
 #' ids <- sample(1:100, 10, replace=FALSE)
 #' ids_alphab <- alphaLabel(ids)
 #' @rawNamespace import(utils)
@@ -12,8 +16,13 @@
 
 alphaLabel <- function(x){
 
+  if(is.character(x)){
+    stop("*== A vector of numbers is required! ==*")
+  }
+
   combind_A <- LETTERS
   combind <-  combn(LETTERS, m=2, sep="")
+
   list_Letters <- NULL
   for(cc in seq_len(ncol(combind))){#cc=1
     list_Letters <-c(list_Letters,  paste(combind[1,cc],
@@ -21,10 +30,10 @@ alphaLabel <- function(x){
   }
   list_Letters <- c(combind_A, list_Letters)
   if(length(unique(x))<=350){
-    clust_num <- list_Letters[x]
+    clust_num <- list(alphabetic_Labels=list_Letters[x])
     return(clust_num)
   }
   if(length(unique(x))>350){
-    print("Labels exhausted! specify a vector with fewer elements")
+    stop("Labels exhausted! specify a vector with fewer elements")
   }
 }
