### 2020-08-18 78c2210
updates props
