### 2020-08-18 c5a9093
update outliers
