### 2020-08-18 3b76c79
updates elbowPoints
