+++ man/dataImputation.Rd
@@ -4,39 +4,68 @@
 \alias{dataImputation}
 \title{Data imputation for longitudinal data}
 \usage{
-dataImputation(traj, id_field = FALSE, method = 2, replace_with = 1, fill_zeros = FALSE)
+dataImputation(traj, id_field = FALSE, method = 2,
+replace_with = 1, fill_zeros = FALSE)
 }
 \arguments{
-\item{traj}{[\code{matrix (numeric)}]: longitudinal data. Each row represents an individual trajectory (of observations). The columns show the observations at consecutive time points.}
+\item{traj}{[\code{matrix (numeric)}]: longitudinal data. Each
+row represents an individual trajectory (of observations). The
+columns show the observations at consecutive time points.}
 
-\item{id_field}{[numeric or character] Whether the first column of the \code{traj} is a unique (\code{id}) field. Default: \code{FALSE}. If \code{TRUE} the function recognises the second column as the first time step.}
+\item{id_field}{[numeric or character] Whether the first column
+of the \code{traj} is a unique (\code{id}) field. Default:
+\code{FALSE}. If \code{TRUE} the function recognises the second
+column as the first time step.}
 
-\item{method}{[an integer] indicating a method for calculating the missing values. Options are: \code{'1'}: \code{arithmetic} method, and \code{'2'}: \code{regression} method. The default is \code{'1'}: \code{arithmetic} method}
+\item{method}{[an integer] indicating a method for calculating
+the missing values. Options are: \code{'1'}: \code{arithmetic}
+ method, and \code{'2'}: \code{regression} method. The default
+ is \code{'1'}: \code{arithmetic} method}
 
-\item{replace_with}{[an integer from 1 to 6] indicating the technique, based on a specified \code{method}, for calculating the missing entries.
-\code{'1'}: \code{arithmetic} method, \code{replace_with} options are: \code{'1'}: Mean value of the corresponding column;
-\code{'2'}: Minimum value of corresponding column; \code{'3'}: Maximum value of corresponding column;
-\code{'4'}: Mean value of corresponding row; \code{'5'}: Minimum value of corresponding row,
-or \code{'6'}: Maximum value of corresponding row. For \code{'2'}: regression method:
-the available option for the \code{replace_with} is: \code{'1'}: \code{linear}.
-The regression method fits a linear regression line to a trajectory with missing entry(s)
+\item{replace_with}{[an integer from 1 to 6] indicating the technique,
+based on a specified \code{method}, for calculating the missing entries.
+\code{'1'}: \code{arithmetic} method, \code{replace_with} options are:
+\code{'1'}: Mean value of the corresponding column;
+\code{'2'}: Minimum value of corresponding column; \code{'3'}:
+Maximum value of corresponding column;
+\code{'4'}: Mean value of corresponding row; \code{'5'}:
+Minimum value of corresponding row,
+or \code{'6'}: Maximum value of corresponding row. For \code{'2'}:
+regression method:
+the available option for the \code{replace_with} is: \code{'1'}:
+\code{linear}.
+The regression method fits a linear regression line to a trajectory
+with missing entry(s)
 and estimates the missing data values from the regression line.
-Note: only the missing data points derive their new values from the regression line
-while the rest of the data points retain their original values. The function terminates if there are
-trajectories with only one observation. The default is \code{'1'}: Mean value of the corresponding column}
+Note: only the missing data points derive their new values from the
+regression line
+while the rest of the data points retain their original values. The
+function terminates if there are
+trajectories with only one observation. The default is \code{'1'}: Mean
+value of the corresponding column}
 
[…]
