### 2020-08-19 #2379 @MAnalytics
Dear @karthik @njtierney

This is to inform you that I have now addressed all the issues raised above.
I know you must be very busy, especially during this time period, I will greatly appreciate it if you can promptly review these corrections so that we can proceed to the next stage.
Thank you very much.
MA

**Summary of revisions:**
1. Included unit tests for all function, and all package codes.
2. Imported only required library functions (instead of whole packages).
3. Change all assignment signs '=' to '<-'
4. Short code line (<80 characters wide) implemented for all functions
5. Used seq_len( ) function to prevent errors associated with 1:nrow( ), 1:length( ), ....
6. 'TRUE' and 'FALSE' utilized instead of 'T' and 'F', respectively.
7. Language field included in the DESCRIPTION file
8. All spelling errors addressed.
