+++ NAMESPACE
@@ -10,7 +10,34 @@ export(rates)
 export(removeRowsN)
 export(statPrint)
 export(wSpaces)
-import(kml, grDevices, reshape2, Hmisc, stats, utils, ggplot2, longitudinalData)
-import(reshape2, ggplot2, stats)
-import(stats)
-import(utils)
+importFrom(Hmisc,cut2)
+importFrom(clusterCrit,intCriteria)
+importFrom(ggplot2,aes)
+importFrom(ggplot2,facet_wrap)
+importFrom(ggplot2,geom_area)
+importFrom(ggplot2,geom_line)
+importFrom(ggplot2,geom_point)
+importFrom(ggplot2,geom_vline)
+importFrom(ggplot2,ggplot)
+importFrom(ggplot2,ggtitle)
+importFrom(ggplot2,scale_colour_brewer)
+importFrom(ggplot2,scale_fill_brewer)
+importFrom(ggplot2,scale_x_continuous)
+importFrom(ggplot2,stat_summary)
+importFrom(ggplot2,theme)
+importFrom(ggplot2,theme_light)
+importFrom(grDevices,dev.new)
+importFrom(kml,affectIndivC)
+importFrom(reshape2,melt)
+importFrom(signal,sgolayfilt)
+importFrom(stats,approx)
+importFrom(stats,approxfun)
+importFrom(stats,coefficients)
+importFrom(stats,lm)
+importFrom(stats,median)
+importFrom(stats,optimize)
+importFrom(stats,predict)
+importFrom(stats,quantile)
+importFrom(stats,smooth.spline)
+importFrom(utils,combn)
+importFrom(utils,flush.console)
