### 2020-08-19 6c50e82
update main clustering function
