### 2020-08-19 b90ece3
import stats functions
