+++ R/dataImputation.R
@@ -46,12 +46,16 @@
 #' corresponding rows or columns, or to use a regression method to
 #' estimate the missing values.
 #' @examples
+#'
 #' print(traj)
+#'
 #' dataImputation(traj, id_field = TRUE, method = 1, replace_with = 1,
 #' fill_zeros = FALSE)
-#' @rawNamespace import(stats)
+#'
 #' @return A data.frame with missing values (\code{NA}, \code{Inf},
 #' \code{null}) imputed according to the a specified technique.
+#' @importFrom stats predict
+#' @importFrom utils flush.console
 #' @export
 
 dataImputation <- function(traj, id_field = FALSE, method = 2,
