### 2020-08-19 a1f7de1
update elbowPoint.R
