### 2020-08-19 32c2fe5
update outlierDetect.R
