### 2020-08-19 c4d7495
adding functions
