### 2020-08-19 618c16d
removeRowsN functions
