### 2020-08-19 30879b7
update statPrints.R
