### 2020-08-19 0969913
update wSpaces.R
