### 2020-08-19 523b075
add unit test for akmedoids.clust.R
