### 2020-08-19 7e2fe48
update alphaLabel.R
