### 2020-08-19 44765e4
elbowPoint.R unit
