+++ tests/testthat/test_elbowPoint.R
@@ -1,3 +1,4 @@
+context("Testing elbowPoint.R function");
 
 x <- runif(100, min=-2, max=3)
 y <- -exp(-x) * (1+rnorm(100)/3)
