### 2020-08-19 a7613dd
update test_rates.R
