### 2020-08-19 e78691f
update test_statPrint.R
