### 2020-08-19 29dee37
update test_traj.R
