### 2020-08-19 8a86761
update test_wSpaces.R
