+++ tests/testthat/test_wSpaces.R
@@ -1,4 +1,4 @@
-context("Tests for checking data types for wSpaces function");
+context("Testing wSpaces.R function");
 
 #simulating different whitespaces
 sn <- c('sn1','sn2','sn3')
