### 2020-08-19 abf4e20
add vignette Rmd
