### 2020-08-19 0db54ce
update akmedoids.Rd
