+++ man/akmedoids.clust.Rd
@@ -4,39 +4,93 @@
 \alias{akmedoids.clust}
 \title{Anchored k-medoids clustering}
 \usage{
-akmedoids.clust(traj, id_field = FALSE, method = "linear", k = c(3,6), crit="Silhouette")
+akmedoids.clust(traj, id_field = FALSE, method = "linear",
+k = c(3,6), crit="Silhouette")
 }
 \arguments{
-\item{traj}{[matrix (numeric)]: longitudinal data. Each row represents an individual trajectory (of observations). The columns show the observations at consecutive time steps.}
+\item{traj}{[matrix (numeric)]: longitudinal data. Each row represents
+an individual trajectory (of observations). The columns show the
+observations at consecutive time steps.}
 
-\item{id_field}{[numeric or character] Whether the first column of the \code{traj} is a unique (\code{id}) field. Default: \code{FALSE}. If \code{TRUE} the function recognizes the second column as the first time points.}
+\item{id_field}{[numeric or character] Whether the first column of the
+\code{traj} is a unique (\code{id}) field. Default: \code{FALSE}.
+If \code{TRUE} the function recognizes the second column as the first
+time points.}
 
-\item{method}{[character] The parametric initialization strategy. Currently, the only available method is a \code{linear} method, set as \code{"linear"}. This uses the time-dependent linear regression lines and the resulting groups are order in the order on increasing slopes.}
+\item{method}{[character] The parametric initialization strategy.
+Currently, the only available method is a \code{linear} method, set as
+\code{"linear"}. This uses the time-dependent linear regression lines
+and the resulting groups are order in the order on increasing slopes.}
 
-\item{k}{[integer or vector (numeric)] either an exact integer number of clusters, or a vector of length two specifying the minimum and maximum numbers of clusters to be examined from which the best solution will be determined. In either case, the minimum number of clusters is \code{3}. The default is \code{c(3,6)}.}
+\item{k}{[integer or vector (numeric)] either an exact integer number
+of clusters, or a vector of length two specifying the minimum and
+maximum numbers of clusters to be examined from which the best
+solution will be determined. In either case, the minimum number
+of clusters is \code{3}. The default is \code{c(3,6)}.}
 
-\item{crit}{[character] a string specifying the type of the criterion to use for assessing the quality of the cluster solutions, when \code{k} is a vector of two values (as above). Default:\code{crit="Silhouette"}, use the average Silhouette width (\code{Rousseeuw P. J. 1987}). Using the \code{"Silhouette"} criterion, the optimal value of \code{k} can be determined as the elbow point of the curve. Other valid criterion is the "Caliński_Harabasz" (\code{Caliński T. & Harabasz J. 1974}) in which the maximum score represents the point of optimality. Having determined the optimal \code{k}, the function can then be re-run, using the exact (optimal) value of \code{k}.}
+\item{crit}{[character] a string specifying the type of the criterion
+to use for assessing the quality of the cluster solutions, when
+\code{k} is a vector of two values (as above). Default:
+\code{crit="Silhouette"}, use the average Silhouette width
+(\code{Rousseeuw P. J. 1987}). Using the \code{"Silhouette"} criterion,
+the optimal value of \code{k} can be determined as the elbow point of
+the curve. Other valid criterion is the "Caliński_Harabasz"
+(\code{Caliński T. & Harabasz J. 1974}) in which the maximum score
+represents the point of optimality. Having determined the optimal
+\code{k}, the function can then be re-run, using the exact (optimal)
+value of \code{k}.}
 }
 \value{
[…]
