### 2020-08-19 818ae22
update alphaLabel.Rd
