### 2020-08-19 94d601f
update elbowPoint.Rd
