### 2020-08-19 4703a59
update outlierDetect.Rd
