+++ man/outlierDetect.Rd
@@ -56,11 +56,17 @@ in which the outlier observation is located. Option is also provided to
 remove the trajectories (containing the outlier) from the data.
 }
 \examples{
+
 traj <- traj
-traj <- dataImputation(traj, id_field=TRUE, method = 1, replace_with = 1)
-traj <- props(traj, id_field=TRUE)#remove this later
-outlierDetect(traj, id_field = TRUE, method = 1, threshold = 0.95,
+
+trajectry <- dataImputation(traj, id_field=TRUE, method = 1, replace_with = 1)
+
+trajectry <- props(trajectry$CompleteData, id_field=TRUE)
+
+outlierDetect(trajectry, id_field = TRUE, method = 1, threshold = 0.95,
 count = 1, replace_with = 1)
-outlierDetect(traj, id_field = TRUE, method = 2, threshold = 15,
+
+outlierDetect(trajectry, id_field = TRUE, method = 2, threshold = 15,
 count = 4, replace_with = 3)
+
 }
