### 2020-08-19 c7ce290
update rates.Rd
