### 2020-08-19 10f56d2
update removeRowsN.Rd
