### 2020-08-19 882e7a9
update statPrint.Rd
