### 2020-08-19 79df303
update wSpaces.Rd
