### 2020-08-19 #2379 @karthik
Some of these issues still remain. Here are a few examples.

https://github.com/MAnalytics/akmedoids/blob/master/R/akmedoids.clust.R#L153

https://github.com/MAnalytics/akmedoids/blob/master/R/akmedoids.clust.R#L167

https://github.com/MAnalytics/akmedoids/blob/master/R/dataImputation.R#L175
