### 2020-08-19 74435ff
update akmedoids.clust.R
