### 2020-08-19 #2379 @MAnalytics
@karthik I really appreciate your prompt review.
**The highlighted issues have now been resolved.** Thanks, MA
