### 2020-08-19 #2379 @karthik
There are still several more of these issues. Please look through your code where you do 1:nrow or 1:ncol. If the dimensions are 0 you will get silent errors that will go uncaught.

https://github.com/MAnalytics/akmedoids/blob/master/R/statPrint.R#L126
https://github.com/MAnalytics/akmedoids/blob/master/R/statPrint.R#L210
https://github.com/MAnalytics/akmedoids/blob/master/R/statPrint.R#L261
https://github.com/MAnalytics/akmedoids/blob/master/R/statPrint.R#L321
