### 2020-08-19 f1d4d73
updating statPrint.R
