### 2020-08-19 26f0043
update rates.Rd
