### 2020-08-19 905e7be
update akmedoid.clust.R
