### 2020-08-19 f2f4cee
updates rates.R
