### 2020-08-19 #2379 @MAnalytics
@karthik I have gone through all the functions and corrected these issues accordingly.
Thanks, MA
