### 2020-08-19 4252c11
updating vignette
