### 2020-08-19 #2379 @njtierney
Running `googpractice::gp()` I found a few missed cases:

I note that `1:length` is used here:

https://github.com/MAnalytics/akmedoids/blob/master/tests/testthat/test_rates.R#L19

FYI There are minor also some style changes that can be fixed

- using = instead of <-
- using T and F instead of TRUE and FALSE
- using trailing `;`

But I can resume my review from here, these are minor things. I would recommend running `googpractice::gp()`
