### 2020-08-20 d60b885
update test_rates.R
