### 2020-08-20 48d7d91
update akmedoids.clust.R
