+++ R/dataImputation.R
@@ -203,7 +203,7 @@ dataImputation <- function(traj, id_field = FALSE, method = 2,
       #add to the original data.
       dat[k, known_1[,1]] <- round(newY, digits = 2)
     }
-    model = "lm(y ~ x, data = data)"
+    model <- "lm(y ~ x, data = data)"
     datF <- dat
   }
 
