### 2020-08-20 d4f2fbb
updating elbowPoint.R
