### 2020-08-20 8a64646
update removeRowsN.R
