### 2020-08-20 33e0b29
update statPrint.R
