### 2020-08-20 2fd2369
update wSpaces.R
