### 2020-08-20 ccad35e
update test_akmedoids.clust.R
