### 2020-08-20 145a7ca
update test_statPrint.R
