### 2020-08-20 4d4d853
update test_wSpaces.R
