### 2020-08-20 f495c6c
updating akmedoids.clust.Rd
