+++ man/akmedoids.clust.Rd
@@ -81,7 +81,8 @@ print(trajectry)
 output <- akmedoids.clust(trajectry, id_field = TRUE,
 method = "linear", k = c(3))
 
-print(output)  #type 'as.vector(output$memberships)'
+print(output)
+#type 'as.vector(clusters$memberships)$alphabetic_Labels'
 
 }
 \references{
