### 2020-08-20 14928e5
update test_akmedoid.clust
