### 2020-08-20 970a14b
Update alphaLable.R
