### 2020-08-20 7fa81fb
test elbowPoint.R
