### 2020-08-20 c3bc5e6
update test_outlierDetect
