### 2020-08-20 8c012c0
update props.R
