### 2020-08-20 3c0f3c2
update test_ratef
