+++ tests/testthat/test_rates.R
@@ -1,4 +1,4 @@
-context("Testing rates.R function");
+context("Testing rates.R function")
 
 #utilise trajectory data ('traj.rda')
 traj2 <- dataImputation(traj, id_field = TRUE, method = 2,
@@ -24,8 +24,8 @@ list_ <- c(2, 3, 5, 6, 7, 9, 10) #vector of missing years
 pop_imp_result <- dataImputation(pop2, id_field = TRUE, method = 2,
   replace_with = 1, fill_zeros = FALSE)
 
-crime_rates <- rates(traj=traj2$CompleteData, denomin=pop_imp_result$CompleteData,
-  multiplier = 200)
+crime_rates <- rates(traj=traj2$CompleteData,
+              denomin=pop_imp_result$CompleteData, multiplier = 200)
 
 test_that('dimension of test data', {
   expect_equal(length(crime_rates), 4)
@@ -61,7 +61,8 @@ test_that('output correct error messages', {
   expect_error(rates(traj=traj2, denomin=pop_imp_result,id_field=FALSE,
    prints_text("*---unique field must be set as 'TRUE'!---*")))
   expect_error(rates(traj=traj2, denomin=population,id_field=TRUE,
-   prints_text("(: The 'id_field' of the 'traj' object is not a unique field. Function terminated!!! :)")))
+   prints_text("(: The 'id_field' of the 'traj' object is not a",
+               "unique field. Function terminated!!! :)", sep=" ")))
 
 })
 
