### 2020-08-20 0f62457
test removeRows.R
