### 2020-08-20 #2379 @MAnalytics
@njtierney Thank you  very much for the suggestion of googpractice::gp()
I have now used the package to fix all the issues.
Thanks.
