### 2020-08-21 673add4
update akmedoid.clust.R
