### 2020-08-21 534f32e
update elbowPoint.R
