### 2020-08-21 08ea709
update props.R
