+++ R/props.R
@@ -14,10 +14,11 @@
 #' @usage props(traj, id_field = TRUE, scale = 1, digits = 4)
 #' @examples
 #'
-#' trajectry <- dataImputation(traj, id_field = TRUE, method = 2, replace_with = 1,
-#' fill_zeros = FALSE) #filling the missing values
+#' trajectry <- dataImputation(traj, id_field = TRUE, method = 2,
+#' replace_with = 1, fill_zeros = FALSE) #filling the missing values
 #'
-#' trajectry <- props(trajectry$CompleteData, id_field = TRUE, scale=1, digits=4)
+#' trajectry <- props(trajectry$CompleteData, id_field = TRUE,
+#' scale=1, digits=4)
 #'
 #' print(trajectry)
 #'
