### 2020-08-21 628aece
update statprint.R
