### 2020-08-21 dd7e841
udpate wSpace.R
