### 2020-08-21 8bdeb0d
update test_akmedoids.clust.R
