### 2020-10-21 #2379 @karthik
@MAnalytics Can you please add these files

```
Non-standard files/directories found at top level:
    'cran-comments.md' 'paper.bib' 'paper.md'
```

to your `..Rbuildignore`

Please fix devtools installation instructions in your vignette. Currently it is:
```
devtools::install_github("manalytics/packages/akmedoids")
```
(remove `packages`).
