### 2020-10-21 #2379 @karthik
Please also cite R in the text "A full demonstration is provided in the package vignette of how to deploy `Akmedoids` to examine long-term relative exposure to crime in `R`"

R citation: https://cran.r-project.org/doc/FAQ/R-FAQ.html#Citing-R
