+++ paper.bib
@@ -133,3 +133,11 @@
   year = 2016
 }
 
+@Manual{rManual:2020,
+    author = {R Core Team},
+    title = {R: A Language and Environment for Statistical Computing},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    url = {https://www.R-project.org/},
+    year = 2020
+  }
