### 2020-10-21 #2379 @MAnalytics
@karthik @njtierney Thanks.
I have now addressed these highlighted issues.
MA.
