### 2020-10-30 #2379 @njtierney
Thank you for your patience everyone, here is my review.

I have a few concerns regarding the functionality of the software and how easy it is to use for a data analysis. I appreciate that perhaps these points might at first appearance appear to be superficial, however I think that making these improvements will make the user experience better, and improve the package overall - I myself would love to use some of this work in my longitudinal data analysis, however at the moment I think that there are some improvements that can be made to the code.

I am very happy to provide more pointers/advice/detail if you have any questions about my review or comments on the code.

Thank you again for your submission, I enjoyed reading it, and appreciate your strong effort in making this project open source.

## Review checklist for [@njtierney]

-   [ ] **Contribution and authorship:** Has the submitting author ([@MAnalytics]) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

    -   @MAnalytics appears to have been a major contributor to the software, although there are a lot of commits from `geoMADE`, but this does not appear to be an account (<https://github.com/geoMADE),> so I assume that these commits are from @MAnalytics and are named as geoMADE somehow.

One common theme I've seen in this package is that there is a lot of code being printed out to the console when functions are being run, and there is no way to silence this. For example:

```{r example-clust}
example_clust <- akmedoids.clust(all_traj, k=c(3,10), crit = "Silhouette")
```

returns

```
[1] "solution of k = 3 determined!"
[1] "solution of k = 4 determined!"
[1] "solution of k = 5 determined!"
[1] "solution of k = 6 determined!"
[1] "solution of k = 7 determined!"
[1] "solution of k = 8 determined!"
[1] "solution of k = 9 determined!"
[1] "solution of k = 10 determined!"
[1] "Suggested optimal solution contains 3 clusters. See the plot for further examination!"
```

These print messages are useful, but there should be a `verbose` or `quiet` argument to suppress these. This information should be also stored in the output of the object.

This function also opened a graphic window and plot. There is no way for the user to suppress this. The output should be stored in the object, and if you really want to show the plot to uses, have an option to print the plot when running the function, perhaps with an argument like `show_plot = TRUE`.

Reading the documentation of the package, the function returns different output based on the value of K. The documentation reads:

> #' @return If \code{k} is a vector of two numbers (see param.
#' \code{k} details above), the output is a graphical plot of
#' the quality scores of the cluster solutions. If \code{k} is
#' an exact integer number of clusters, the function returns
#' trajectory labels indicating the group membership of the
#' corresponding trajectory in the \code{traj} object.

I think that this makes the software harder to use, as there is an implied workflow that could be made simpler and easier to use.

It seems then that the next step is to run the same code again but with `k` = 3. I am not sure why this needs to be rerun, shouldn't this be stored as the output? For example, a vector of classes for each value of `k`.

Running the function again with `k = 3`, I get the following:

```{r example-clust-3}
example_clust_3 <- akmedoids.clust(all_traj, k=c(3), crit = "Silhouette")
example_clust_3
```

Finally to summarise the information we run:

#get the optimal cluster solution
clustr <- akmedoids.clust(all_traj, k=3)
clustr <- as.vector(clustr$memberships)$alphabetic_Labels

##get group statistics
descrStats <- statPrint(clustr, all_traj,
                       id_field = FALSE, showplots=FALSE)

print(descrStats)
```

I believe that this can be improved in a few ways.

Firstly, writing

```r
clustr <- akmedoids.clust(all_traj, k=3)
clustr <- as.vector(clustr$memberships)$alphabetic_Labels
```

should be avoided as you are overwriting the object.

Secondly, the user should not need to perform special data manipulation (such as accessing a nested list with (`as.vector(clustr$memberships)$alphabetic_Labels`) in order to summarise the data.

This could be improved by writing a special `print` method for the software. This would mean that  printing `clustr` out to the console, e.g.,

```r
clustr <- akmedoids.clust(all_traj, k=3)
clustr
```

Outputting the `clustr` object in the console, it could return the information that `statPrint` returns. You can read more about creating custom `print` methods here - https://adv-r.hadley.nz/s3.html, https://blog.earo.me/2019/11/03/practical-guide-to-s3/.

It seems that there is a workflow implied:

1. Test multiple different values of `k` with `akmedoids.clust`
  - explore the plots of the optimal solution based on a certain criterion.
  - decide on ideal value of K
2. Run clustering for the optimally identified value of `k`
  - using `akmedoids.clust` again
3. summarise this information using `statPrint`

I think that this could be improved in the following way.

1. Provide a function for exploring different values of K
2. Provide a function for identifying the cluster labels
3. Provide good print methods, so the output that is given to the console could be delivered via the print method.

I suggestion is the output of this function should contain possible memberships for

I also note that `statPrint` is noted as performing two tasks - (1) generating statistics of groups, and (2), creating plots.

In my opinion the output of `statPrint` and `akmedoids.clust` functions should be separated into two functions, or at least provide options for the plots to be turned off. If they are turned into functions, I recommend:

- One function for identifying clusters or summary statistics that return a data frame output or a list.
- And one function that focuses on creating graphics.

I have some other comments on the functions in this package:

- `akmedoids::alphaLabel()`
  - I'm not sure if the output of this needs to be a list, it could instead just be the vector of characters.
  - Is it possible to provide an example of the use of this function?

- `akmedoids::dataImputation()`
  - see note in "common issues" about printing messages

- `akmedoids::elbowPoint()`
  - In example documentation, the function is applied to the plot (which is great!) but is not demonstrated in other contexts. I suggest providing an example outside of plotting as well.

`akmedoids::outlierDetect()`
  - This prints out several messages and objects, see below note on messages.

- `akmedoids::props()`
  - states output is a matrix, but is a data.frame

- `akmedoids::rates()`
  - This function doesn't return a matrix, instead returning a vector of common IDs. I am not sure what these mean, as their behaviour is not documented in the helpfile.

- `akmedoids::removeRowsN()`
  - Could the documentation clarify how is this different to `na.omit()`?

- `akmedoids::wSpaces()`
  - The example for this function doesn't contain whitespace characters
  - Could the documentation explain how this is different to `trimws` function in R?

- `akmedoids::statPrint()`

  - It is very neat that this function can generate a nice looking ggplot, however it would be great if that plot was instead a separate function

#### Common issues

- plots all open new graphics devices rather than using standard `print` methods, which can be problematic when trying to include them in documents
- no way to silence console output - suggest adding a `quiet` or `verbose` option, and to use `message` when printing messages to the console.
  - e.g., the `rates`, `akmedoids.clust` functions

> Yes.

The statement in the README says:

>  identify clusters of trajectories with similar long-term linear trends over time, providing an improved cluster identification as compared with the classic kmeans algorithm

Yes

> Although the author provides instructions in the README, but these contain no output of the function. This should be updated to make the README a `README.Rmd` function. See `usethis::use_readme_rmd()` as an approach on doing this.

> In addition, the README simulates some example data for use - this should be encapsulated as an example dataset in the package that can be loaded.

- example documentation has unnecessary `traj <- traj` line in "examples"
- The datasets are not described in the helpfiles, and state that they are matrices when they are data frames. This data should be documented more thoroughly (see https://r-pkgs.org/data.html#documenting-data for details on how to document data, and https://github.com/jforbes14/eechidna/blob/master/R/data.R for an example documentation of data).

> There are tests
