+++ R/akmedoids.clust.R
@@ -1,446 +1,0 @@
-
-#' @title Anchored k-medoids clustering
-#' @description Given a list of trajectories and a functional method,
-#' this function clusters the trajectories into a \code{k} number of
-#' groups. If a vector of two numbers is given, the function determines
-#' the best solution from those options based on the Caliński-Harabasz
-#' criterion.
-#' @param traj [matrix (numeric)]: longitudinal data. Each row represents
-#' an individual trajectory (of observations). The columns show the
-#' observations at consecutive time steps.
-#' @param id_field [numeric or character] Whether the first column of the
-#' \code{traj} is a unique (\code{id}) field. Default: \code{FALSE}.
-#' If \code{TRUE} the function recognizes the second column as the first
-#' time points.
-#' @param method [character] The parametric initialization strategy.
-#' Currently, the only available method is a \code{linear} method, set as
-#' \code{"linear"}. This uses the time-dependent linear regression lines
-#' and the resulting groups are order in the order on increasing slopes.
-#' @param k [integer or vector (numeric)] either an exact integer number
-#' of clusters, or a vector of length two specifying the minimum and
-#' maximum numbers of clusters to be examined from which the best
-#' solution will be determined. In either case, the minimum number
-#' of clusters is \code{3}. The default is \code{c(3,6)}.
-#' @param crit [character] a string specifying the type of the criterion
-#' to use for assessing the quality of the cluster solutions, when
-#' \code{k} is a vector of two values (as above). Default:
-#' \code{crit="Silhouette"}, use the average Silhouette width
-#' (\code{Rousseeuw P. J. 1987}). Using the \code{"Silhouette"} criterion,
-#' the optimal value of \code{k} can be determined as the elbow point of
-#' the curve. Other valid criterion is the "Caliński_Harabasz"
-#' (\code{Caliński T. & Harabasz J. 1974}) in which the maximum score
-#' represents the point of optimality. Having determined the optimal
-#' \code{k}, the function can then be re-run, using the exact (optimal)
-#' value of \code{k}.
-#' @usage akmedoids.clust(traj, id_field = FALSE, method = "linear",
-#' k = c(3,6), crit="Silhouette")
-#' @details This function works by first approximating the trajectories
-#' based on the chosen parametric forms (e.g. linear), and then partitions
-#' the original trajectories based on the form groupings, in similar
-#' fashion to k-means clustering \code{(Genolini et al. 2015)}. The key
-#' distinction of \code{akmedoids} compared with existing longitudinal
-#' approaches is that both the initial starting points as well as the
-#' subsequent cluster centers (as the iteration progresses) are based
-#' the selection of observations (medoids) as oppose to centroids.
-#' @examples
-#'
-#' traj <- traj
-#'
-#' print(traj)
-#'
-#' trajectry <- dataImputation(traj, id_field = TRUE, method = 2,
-#' replace_with = 1, fill_zeros = FALSE)
-#'
-#' trajectry <- props(trajectry$CompleteData, id_field = TRUE)
-#'
-#' print(trajectry)
-#'
-#' output <- akmedoids.clust(trajectry, id_field = TRUE,
-#' method = "linear", k = c(3))
-#'
-#' print(output)
-#' #type 'as.vector(clusters$memberships)$alphabetic_Labels'
-#'
-#' @return If \code{k} is a vector of two numbers (see param.
-#' \code{k} details above), the output is a graphical plot of
[…]

[…]
