+++ R/traj_wSpaces.R
@@ -1,0 +1,9 @@
+#' @title Datasets with white spaces
+#' @description Trajectories containing white spaces
+#' @name traj_wSpaces
+#' @docType data
+#' @usage traj_wSpaces
+#' @format A dataframe
+#' @keywords datasets
+
+NULL
