+++ R/TO1Risk.R
@@ -1,0 +1,15 @@
+#' @title Time-at-risk for the Adjudicated Toronto
+#' Youth Data (Sample 1)
+#' @description Time-at-risk per year for 378 individuals
+#' in the Toronto, Ontario, Canada.
+#' @name TO1Risk
+#' @docType data
+#' @usage TO1Risk
+#' @references Nielsen, J. (2018) crimCV:
+#' Group-Based Modelling of Longitudinal Data.
+#' R package version 0.9.6.
+#' URL https://CRAN.R-project.org/package=crimCV.
+#' @format A matrix
+#' @keywords datasets
+
+NULL

[…]
