+++ man/print_akStats.Rd
@@ -1,36 +1,27 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/statPrint.R
-\name{statPrint}
-\alias{statPrint}
-\title{Descriptive (Change) statistics and plots}
+% Please edit documentation in R/print_akStats.R
+\name{print_akStats}
+\alias{print_akStats}
+\title{Descriptive (Change) statistics}
 \usage{
-statPrint(
-  clustr,
-  traj,
-  id_field = TRUE,
+print_akStats(
+  akObject,
+  k = 3,
   reference = 1,
   N.quant = 4,
-  showplots = TRUE,
+  show_Plots = TRUE,
   type = "lines",
   y.scaling = "fixed"
 )
 }
 \arguments{
-\item{clustr}{[vector (character)] A vector of cluster
-membership (labels). For instance, the result extracted
-from the \code{\link{akmedoids.clust}} function.}
+\item{akObject}{An output of \code{\link{akClustr}} function.
+The object contains individual trajectories and their cluster
+solution(s) at the specified values of \code{k}. Also, includes
+the optimal value of \code{k} based on the criterion specified.
+at (different) values of \code{k} the \code{traj}.}
 
-\item{traj}{[matrix (numeric)]: corresponding longitudinal
-data used to generate \code{clustr} (with rows
-corresponding to each label of \code{clustr}).
-For example, the first label of \code{clustr}
-is the group label of the first row of
-\code{traj} matrix, and so on.}
-
-\item{id_field}{[numeric or character] Whether the first
-column of the \code{traj} is a unique (\code{id}) field.
-Default: \code{FALSE}. If \code{TRUE} the function
-recognises the second column as the first time step.}
+\item{k}{[integer] \code{k} cluster to generate its solution.}
 
 \item{reference}{[numeric] Specifying the reference line from
 which the direction of each group is measured. Options are:
@@ -48,7 +39,7 @@ This result can be used to further categorize the groups into
 'classes'. For example, groups that fall within the \code{1st}
 quartile may be classified as 'Stable' groups (Adepeju et al. 2019).}
 
-\item{showplots}{[TRUE or FALSE] To display cluster plots.
+\item{show_Plots}{[TRUE or FALSE] To display cluster plots.
 Defaults \code{TRUE}}
 
 \item{type}{[character] plot type. Available options are:
@@ -67,24 +58,23 @@ and statistics.
 This function perform two tasks:
 (i) it generate the descriptive and change statistics
 of groups, particularly suited for the outputs form
-the \code{\link{akmedoids.clust}} function, and
+the \code{\link{akClustr}} function, and
 (ii) generates the plots of the groups (performances).
 }
 \details{
 Generates the descriptive and change statistics
-of the trajectory groupings. Given a vector of group
-membership (labels) and the corresponding data matrix
-(or data.frame) indexed in the same order, this
-function generates all the descriptive and
-change statistics of all the groups.
-The function can generate a line and an area stacked plot
-drawing from the functionalities of the \code{ggplot2} library.
+of the trajectory groupings. Given an \code{akObject} class (from
+the \code{akClustr} function), this function generates both the
+descriptive and change statistics of the clusters.
+There is also an option to show the plots of cluster groups.
+The function draw from the functionalities of the
+\code{ggplot2} library.
 For a more customized visualisation, we recommend that users
 deploy \code{ggplot2} directly (\code{Wickham H. (2016)}).
 }
 \examples{
 
[…]
