+++ README.Rmd
@@ -1,0 +1,182 @@
+---
+#output: github_document
+output: html_document
+---
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "man/figures/README-",
+  out.width = "100%"
+)
+```
+
+# akmedoids
+
+<!-- badges: start -->
+<!-- badges: end -->
+
+An R package for analyzing and clustering longitudinal data
+
+### Description
+
+The `akmedoids` package advances the clustering of longitudinal datasets in order to identify clusters of trajectories with similar long-term linear trends over time, providing an improved cluster identification as compared with the classic kmeans algorithm. The package also includes a set of functions for addressing common data issues, such as missing entries and outliers, prior to conducting advance longitudinal data analysis. One of the key objectives of this package is to facilitate easy replication of a recent paper which examined small area inequality in the crime drop (Adepeju et al. 2020). Many of the functions provided in the `akmedoids` package may be applied to longitudinal data in general. 
+
+**For more information and usability, check out details on [CRAN](https://cran.r-project.org/web/packages/akmedoids/index.html).**
+
+
+### Installation from `CRAN`
+
+From an R console, type:
+
+```{r, echo=TRUE, message=FALSE, eval=TRUE}
+#install.packages("akmedoids")
+library(akmedoids)
+
+#Other libraries
+library(tidyr)
+library(ggplot2)
+library(reshape)
+library(readr)
+```
+To install the development version of the package, type `remotes::install_github("MAnalytics/akmedoids")`. Please, report any installation problems in the issues.
+
+### Example usage:
+
+Given a longitudinal datasets, the following is an example of how `akmedoids` could be used to extract clusters of trajectories with similar long-term trends over time. We will use a simulated dataset (named `simulated.rda`) stored in the `data/` directory.
+
+
+### Generating artificial dataset
+
+Simulating data set which comprised of three clusters with distinct mean directional change over time. Each group contains 50 trajectories.
+
+
+```{r, echo=TRUE, message=FALSE, eval=TRUE}
+dir.create("input") # create a folder
+
+#function for creating longitudinal noise
+noise_fn <- function(x=3, time){
+  rnorm(length(time), mean=0, x)}
+
+#function for simulating a trajectory group
+sim_group <- function(gr_baseline, sd, time){
+  intcp_errors <- rgamma(1, shape=2, scale=sd) #intercept error
+  mean_traj = gr_baseline + intcp_errors
+  traj = mean_traj + noise_fn(intcp_errors, time)
+}
+
+#time steps
+t_steps <- c(0:20)
+
+#increasing group
+i_gr <- NULL
+for(i in seq_len(50)){
+  i_gr <- rbind(i_gr,
+                sim_group(gr_baseline=(0.5*t_steps),
+                          sd=1, time=t_steps))
+}
+
+#stable group
+s_gr <- NULL
+for(i in seq_len(50)){
+  s_gr <- rbind(s_gr,
+                sim_group(gr_baseline=rep(3,length(t_steps)),
+                          sd=1, time=t_steps))
+}
+
+#decreasing group
+d_gr <- NULL
+for(i in seq_len(50)){
+  d_gr <- rbind(d_gr,
+                sim_group(gr_baseline=(10 - (0.5*t_steps)),
+                          sd=1, time=t_steps))
+}
+
+#combine groups
+simulated <- data.frame(rbind(i_gr, s_gr, d_gr))
+
+#add group label
[…]
