+++ NAMESPACE
@@ -1,14 +1,15 @@
 # Generated by roxygen2: do not edit by hand
 
-export(akmedoids.clust)
+S3method(print_akStats,default)
+export(akClustr)
 export(alphaLabel)
 export(dataImputation)
 export(elbowPoint)
 export(outlierDetect)
+export(print_akStats)
 export(props)
 export(rates)
 export(removeRowsN)
-export(statPrint)
 export(wSpaces)
 importFrom(Hmisc,cut2)
 importFrom(clusterCrit,intCriteria)
@@ -17,6 +18,7 @@ importFrom(ggplot2,facet_wrap)
 importFrom(ggplot2,geom_area)
 importFrom(ggplot2,geom_line)
 importFrom(ggplot2,geom_point)
+importFrom(ggplot2,geom_smooth)
 importFrom(ggplot2,geom_vline)
 importFrom(ggplot2,ggplot)
 importFrom(ggplot2,ggtitle)
