+++ README.Rmd
@@ -146,23 +146,112 @@ ggplot(simulated_long, aes(x = Time, y = value, group=ID, color=Groups)) +
 
 ### Performing clustering using `akmedoids`
 
-Performing clustering analysis using `akmedoids` package.
+The `akmedoids::akClustr()` is used to perform the clustering analysis. Using the simulated data, we write:
 
 ```{r, echo=TRUE, message=TRUE, eval=TRUE}
 
 output <- akClustr(Import_simulated, id_field=TRUE, verbose = FALSE, k=c(3,12), crit = "Silhouette",
-                  quality_Plot=TRUE)
+                  quality_Plot=FALSE)
 
-#Remark: The entire output can be printed out to the console by typing `output`
+#RemarkS: The results can be printed out to the console by typing `output`
 
-#To check the optimal cluster size, type:
-output$optimal_k
-
-#Show quality plot
-plot(output$QltyPlot)
-
-#user may decide the examine the plot post clustering....
 ```
+
+In order to preview all the variables of the generated results, type `names(output)`, e.g.
+
+```{r, echo=TRUE, message=TRUE, eval=TRUE}
+
+names(output)
+
+```
+
+* The output variables can be described as follow:
+
+  + `traj` - returns the input data set used for the clustering.
+
+  + `id_field` - indicates whether the input data set included the id field.
+
+  + `solutions` - the list of cluster solutions by `k` values.
+
+  + `qualitycriterion` - the quality criterion specified.
+
+  + `optimal_k` - the optimal value of `k` as determined by the quality criterion.
+
+  + `qualityCrit.List` - the estimated quality of cluster solutions by `k` values.
+
+  + `QltyPlot` - the plot of `qualityCrit.List`, with a red vertical line to indicate the optimal value of `k`.
+
+
+*Accessing the optimal solution*
+
+The plot of `qualityCrit.List` can be displayed by setting the `quality_Plot` argument as `TRUE`. Otherwise, a user may access the plot by printing the variable `output$QltyPlot` to the console. For example,
+
+
+```{r, fig.width=7, fig.height=5, fig.fullwidth=TRUE}
+
+output$QltyPlot
+
+#the intersection of the red line with the x-axis indicates the optimal k value.
+```
+
+The optimal solution is determined to be 3, as also stored by the variable `output$optimal_k`.
+
+
+### Printing Cluster Stats.
+
+The properties (i.e. the descriptive and change statistics) of any of the cluster solutions (i.e. at any value of `k`) can be accessed using the special `print` function `print_akStats`. The print function takes the `akObject` class generated by the `akmedoids::akClustr` function as input as follows: 
+
+```{r, echo=TRUE, message=FALSE, eval=TRUE}
+
+#Specifying the optimal solution, output$optimal_k (i.e. `k = 3`)
+
+prpties = print_akStats(output, k = , type="lines", y.scaling="fixed", show_Plots = FALSE)
+
+```
+Accessing the variables of `prpties`:
+
+```{r, echo=TRUE, message=TRUE, eval=TRUE}
+
+names(prpties)
+
+```
+
+* output Variables described as follow:
+
+  + `descriptiveStats` - descriptive statistics of the cluster.
+  
+  + `changeStats` - change statistics of the cluster
+  
+  + `cluster_Plot` - the line plot or stacked histogram of cluster groups.
+
+### Printing the group properties
+
+*a. Descriptive Stats.*:
+
+```{r, echo=TRUE, message=TRUE, eval=TRUE}
+
+prpties$descriptiveStats
+
+```
+
+*b. Change Stats.*:
+
[…]
