+++ R/popl.R
@@ -1,0 +1,9 @@
+#' @title sample population (denominator) data
+#' @description simulated denominator data for two consecutive census years
+#' @name popl
+#' @docType data
+#' @usage popl
+#' @format A matrix
+#' @keywords datasets
+
+NULL
