+++ vignettes/akmedoids-vignette.Rmd
@@ -180,7 +180,9 @@ library(akmedoids)
 
 ```{r, eval=TRUE}
 
-#viewing the first 6 rows of 'traj' object
+#import and preview the first 6 rows of 'traj' object
+data(traj)
+
 head(traj)
 
 #no. of rows
@@ -266,12 +268,15 @@ The key step towards using the function for this purpose is to create a matrix, 
 
 ```{r, eval=TRUE}
 
-#viewing the data first 6 rows
-head(population)
+#import population data
+data(popl)
 
-nrow(population) #no. of rows
+#preview the data
+head(popl)
 
-ncol(population) #no. of columns
+nrow(popl) #no. of rows
+
+ncol(popl) #no. of columns
 ```
 
 
@@ -280,11 +285,11 @@ The corresponding `input` dataset is prepared as follows and saved as `populatio
 
 ```{r, echo=FALSE}
 #create a matrix of the same rows and column as the `traj` data
-pop <- as.data.frame(matrix(0, nrow(population), ncol(traj)))
+pop <- as.data.frame(matrix(0, nrow(popl), ncol(traj)))
 colnames(pop) <- names(traj) 
-pop[,1] <- as.vector(as.character(population[,1]))
-pop[,4] <- as.vector(as.character(population[,2]))
-pop[,8] <- as.vector(as.character(population[,3]))
+pop[,1] <- as.vector(as.character(popl[,1]))
+pop[,4] <- as.vector(as.character(popl[,2]))
+pop[,8] <- as.vector(as.character(popl[,3]))
 list_ <- c(2, 3, 5, 6, 7, 9, 10)
 for(u_ in seq_len(length(list_))){ #u_<-1
   pop[,list_[u_]] <- "NA"
@@ -428,11 +433,11 @@ Please see the `akmedoids` user manual for the remaining `data manipulation` fun
 
 #2. Data Clustering
 
-Table 2 shows the two main functions required to carry out the longitudinal clustering and generate the descriptive statistics of the resulting groups. The relevant functions are `akClustr` and `statPrint`. The `akClustr` function clusters trajectories according to the similarities of their long-term trends, while the `statPrint` function extracts descriptive and change statistics for each of the clusters. The latter also generates `quality` plots for the best cluster solution.
+Table 2 shows the two main functions required to carry out the longitudinal clustering and generate the descriptive statistics of the resulting groups. The relevant functions are `akClustr` and `print_akStats`. The `akClustr` function clusters trajectories according to the similarities of their long-term trends, while the `print_akStats` function extracts descriptive and change statistics for each of the clusters. The former also generates `quality` plots for the best cluster solution.
 
-The long-term trends of trajectories are defined in terms of a set of OLS regression lines. This allows the clustering function to classify the final groupings in terms of their slopes as `rising`, `stable`, and `falling`. The key benefits of this implementation is that it allows the clustering process to ignore the short-term fluctuations of actual trajectories and focus on their long-term linear trends. Adepeju and colleagues (2019) applied this technique in crime concentration research for measuring long-term inequalities in the exposure to crime at find-grained spatial scales. 
+The long-term trends of trajectories are defined in terms of a set of OLS regression lines. This allows the clustering function to classify the final groupings in terms of their slopes as `rising`, `stable`, and `falling`. The key benefits of this implementation is that it allows the clustering process to ignore the short-term fluctuations of actual trajectories and focus on their long-term linear trends. Adepeju and colleagues (2020) applied this technique in crime concentration research for measuring long-term inequalities in the exposure to crime at find-grained spatial scales. 
 
[…]
