+++ man/popl.Rd
@@ -1,14 +1,14 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/population.R
+% Please edit documentation in R/popl.R
 \docType{data}
-\name{population}
-\alias{population}
+\name{popl}
+\alias{popl}
 \title{sample population (denominator) data}
 \format{
 A matrix
 }
 \usage{
-population
+popl
 }
 \description{
 simulated denominator data for two consecutive census years
