+++ README.Rmd
@@ -199,13 +199,13 @@ The optimal solution is determined to be 3, as also stored by the variable `outp
 
 ### Printing Cluster Stats.
 
-The properties (i.e. the descriptive and change statistics) of any of the cluster solutions (i.e. at any value of `k`) can be accessed using the special `print` function `print_akStats`. The print function takes the `akObject` class generated by the `akmedoids::akClustr` function as input as follows: 
+The properties (i.e. the descriptive and change statistics) of any of the cluster solutions (i.e. for any value of `k`) can be accessed using the special `print` function `print_akStats`. The print function takes the `akObject` class generated by the `akmedoids::akClustr` function as input as follows: 
 
 ```{r, echo=TRUE, message=FALSE, eval=TRUE}
 
 #Specifying the optimal solution, output$optimal_k (i.e. `k = 3`)
 
-prpties = print_akStats(output, k = , type="lines", y.scaling="fixed", show_Plots = FALSE)
+prpties = print_akStats(output, k = 3, type="lines", y.scaling="fixed", show_Plots = FALSE)
 
 ```
 Accessing the variables of `prpties`:
