### 2020-11-15 64da058
update rate.R
