+++ R/data.R
@@ -1,0 +1,43 @@
+#' Sample labels of a cluster group
+#'
+#' A vector of alphabetical labels representing
+#' the optimal solution of `traj` dataset based on `akClust`
+#' function
+#'
+#' @format A vector containing 9 alphabetical labels
+#'     }
+"clustr"
+
+
+#' Simulated population data.
+#'
+#' Sample simulated population data to be used as
+#' the denominator variable against `traj` dataset. Contains
+#' data for two consecutive census years
+#'
+#' @format A vector containing 9 alphabetical labels
+#' \itemize{
+#'   \item location_id: Character id of sample census unit
+#'   at which the population is obtained.
+#'   \item census_2003: Population estimates at the sample
+#'   locations for the census year 2003.
+#'   \item census_2007: Population estimates at the sample
+#'   locations for the census year 2007.
+
+#'     }
+"popl"
+
+#' Simulated longitudinal dataset
+#'
+#' Contains simulated trajectories belonging to one of the
+#' three pre-defined groups, namely (a) decreasing, (b) stable
+#' and (c) increasing groups.
+#'
+#' @format A vector containing 9 alphabetical labels
+#' \itemize{
+#'   \item X1:
+#'   \item X2:
+#'   \item X3:
+
+#'     }
+"popl"
