### 2020-11-15 d433b11
delete data R code from main R directory
