+++ R/simulated.R
@@ -1,10 +1,0 @@
-#' @title Simulated longitudinal dataset
-#' @description Simulated longitudinal datasets containing
-#' pre-defined clusters.
-#' @name simulated
-#' @docType data
-#' @usage simulated
-#' @format A data.frame
-#' @keywords datasets
-
-NULL
