+++ vignettes/references.bib
@@ -1,8 +1,8 @@
 @article{Adepeju2019,
   title = {Anchored k-medoids: A novel adaptation of k-means further refined to measure inequality in the exposure to crime across micro places},
   author = {M. Adepeju and S. Langton and J. Bannister},
-  year = {2019},
-  journal={submitted}
+  year = {2020},
+  journal={Revised}
   },
 
   @article{Andresen2017,
