+++ tests/testthat/test_clustr.R
@@ -1,6 +1,6 @@
 context("Testing clustr.R function")
 
 test_that('data types, dimension of test data', {
-  expect_is(clustr,'character')
-  expect_equal(length(clustr), 9)
+  expect_is(clustr,'data.frame')
+  expect_equal(nrow(clustr), 9)
 })
