### 2020-11-15 c2567f6
update outlierdetect.R
