+++ R/dataImputation.R
@@ -51,7 +51,8 @@
 #'
 #' # Using the example 'traj' datasets
 #'
-#' imp_data <- dataImputation(traj, id_field = TRUE, method = 2, replace_with = 1,
+#' imp_data <- dataImputation(traj, id_field = TRUE, method = 2,
+#' replace_with = 1,
 #' fill_zeros = FALSE, verbose=FALSE)
 #'
 #' @return A data.frame with missing values (\code{NA}, \code{Inf},
