+++ R/data.R
@@ -1,10 +1,12 @@
-#' Sample labels of a cluster group
+#' Sample labels of cluster groups
 #'
-#' A vector of alphabetical labels representing
+#' A dataframe of alphabetical labels representing
 #' the optimal solution of `traj` dataset based on `akClust`
 #' function
 #'
-#' @format A vector containing 9 alphabetical labels
+#' @format A dataframe containing one variable:
+#' \itemize{
+#'   \item label: alphabetical label of clusters
 #'     }
 "clustr"
 
@@ -15,7 +17,7 @@
 #' the denominator variable against `traj` dataset. Contains
 #' data for two consecutive census years
 #'
-#' @format A vector containing 9 alphabetical labels
+#' @format A dataframe with the following variables:
 #' \itemize{
 #'   \item location_id: Character id of sample census unit
 #'   at which the population is obtained.
@@ -23,7 +25,6 @@
 #'   locations for the census year 2003.
 #'   \item census_2007: Population estimates at the sample
 #'   locations for the census year 2007.
-
 #'     }
 "popl"
 
@@ -33,11 +34,121 @@
 #' three pre-defined groups, namely (a) decreasing, (b) stable
 #' and (c) increasing groups.
 #'
-#' @format A vector containing 9 alphabetical labels
+#' @format A dataframe with the following variables:
 #' \itemize{
-#'   \item X1:
-#'   \item X2:
-#'   \item X3:
-
+#'   \item X1: Values across locations at time step 1
+#'   \item X2: Values across locations at time step 2
+#'   \item X3: Values across locations at time step 3
+#'   \item X4: Values across locations at time step 4
+#'   \item X5: Values across locations at time step 5
+#'   \item X6: Values across locations at time step 6
+#'   \item X7: Values across locations at time step 7
+#'   \item X8: Values across locations at time step 8
+#'   \item X9: Values across locations at time step 9
+#'   \item X10: Values across locations at time step 10
+#'   \item X11: Values across locations at time step 11
+#'   \item X12: Values across locations at time step 12
+#'   \item X13: Values across locations at time step 13
+#'   \item X14: Values across locations at time step 14
+#'   \item X15: Values across locations at time step 15
+#'   \item X16: Values across locations at time step 16
+#'   \item X17: Values across locations at time step 17
+#'   \item X18: Values across locations at time step 18
+#'   \item X19: Values across locations at time step 19
+#'   \item X20: Values across locations at time step 20
+#'   \item X21: Values across locations at time step 21
 #'     }
-"popl"
+"simulated"
+
+
+#' Time-at-risk for the Adjudicated Toronto
+#' Youth Data (Sample 1)
+#'
+#' Real-life time-at-risk per year for 378 individuals
+#' from the age of 8 to 38 in the Toronto, Ontario, Canada.
+#' The data is obtained through the R package `crimCV`.
+#' For further information, please see: Nielsen, J. (2018)
+#' crimCV: Group-Based Modelling of Longitudinal Data.
+#' R package version 0.9.6.
+#' URL https://CRAN.R-project.org/package=crimCV.
+#'
+#' @format A dataframe with the following variables:
+#'   \itemize{
+#'   \item 8: Time-at-risk per year at age 8
+#'   \item 9: Time-at-risk per year at age 9
+#'   \item 10: Time-at-risk per year at age 10
+#'   \item 11: Time-at-risk per year at age 11
+#'   \item 12: Time-at-risk per year at age 12
[…]
