### 2020-11-15 2340963
update akClustr.R
