+++ man/popl.Rd
@@ -1,16 +1,26 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/popl.R
+% Please edit documentation in R/data.R
 \docType{data}
 \name{popl}
 \alias{popl}
-\title{sample population (denominator) data}
+\title{Simulated population data.}
 \format{
-A matrix
+A dataframe with the following variables:
+\itemize{
+  \item location_id: Character id of sample census unit
+  at which the population is obtained.
+  \item census_2003: Population estimates at the sample
+  locations for the census year 2003.
+  \item census_2007: Population estimates at the sample
+  locations for the census year 2007.
+    }
 }
 \usage{
 popl
 }
 \description{
-simulated denominator data for two consecutive census years
+Sample simulated population data to be used as
+the denominator variable against `traj` dataset. Contains
+data for two consecutive census years
 }
 \keyword{datasets}
