+++ man/clustr.Rd
@@ -1,16 +1,21 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/clustr.R
+% Please edit documentation in R/data.R
 \docType{data}
 \name{clustr}
 \alias{clustr}
-\title{Sample clusters}
+\title{Sample labels of cluster groups}
 \format{
-A vector
+A dataframe containing one variable:
+\itemize{
+  \item label: alphabetical label of clusters
+    }
 }
 \usage{
 clustr
 }
 \description{
-A vector of cluster labels
+A dataframe of alphabetical labels representing
+the optimal solution of `traj` dataset based on `akClust`
+function
 }
 \keyword{datasets}
