### 2020-11-15 d7a1631
update data.R
