### 2020-11-16 1aaa923
update print_akStats.R
