### 2020-11-16 b144ae6
update removeRowsN.R
