### 2020-11-16 ce3161b
update manuals
