+++ man/removeRowsN.Rd
@@ -25,11 +25,9 @@ A matrix with complete observations
 }
 \description{
 This function removes any rows in which an 'NA'
-or an 'Inf' entry is found.The ability to remove records with
-'Inf' entries distinguishes this function from the popular
-'na.omit()' function.
- removal of '' function differs from
-'na.omit(data)'
+or an 'Inf' entry is found. The function is also able to
+remove records with 'Inf' entries, distinguishing it from
+the popular 'na.omit()' function in R.
 }
 \details{
 Given a matrix (or a dataframe) containing an 'NA' or
