+++ man/clustr.Rd
@@ -7,7 +7,7 @@
 \format{
 A dataframe containing one variable:
 \itemize{
-  \item label: alphabetical label of clusters
+  \item label: alphabetical label by clusters
     }
 }
 \usage{
