### 2020-11-16 #2379 @MAnalytics
Dear @njtierney (and @karthik),

We greatly appreciate you taking the time to review our paper.

We have addressed all the issues raised by @njtierney, and we hope that the paper will now be accepted for publication. Please, find our response to each of the issues below.

Kind Regards,

Monsuru (@MAnalytics)

***#-----------------------------------***

> Thank you for your patience everyone, here is my review.
>
> I have a few concerns regarding the functionality of the software and how easy it is to use for a data analysis. I appreciate that perhaps these points might at first appearance appear to be superficial, however I think that making these improvements will make the user experience better, and improve the package overall - I myself would love to use some of this work in my longitudinal data analysis, however at the moment I think that there are some improvements that can be made to the code.
>
> I am very happy to provide more pointers/advice/detail if you have any questions about my review or comments on the code.
>
> Thank you again for your submission, I enjoyed reading it, and appreciate your strong effort in making this project open source.
>
> ## Review checklist for [@njtierney]
> ### Conflict of interest
> * [x]  I confirm that I have read the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/master/COI.md) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
> ### Code of Conduct
> * [x]  I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [x]  **Repository:** Is the source code for this software available at the [repository url](https://github.com/MAnalytics/akmedoids)?
> * [x]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
> * [ ]  **Contribution and authorship:** Has the submitting author ([@MAnalytics]) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
>
***## ------------------------------------------***
***1. Reviewer's comment:***
***## ------------------------------------------***
>
>   * @MAnalytics appears to have been a major contributor to the software, although there are a lot of commits from `geoMADE`, but this does not appear to be an account (https://github.com/geoMADE), so I assume that these commits are from @MAnalytics and are named as geoMADE somehow.
>
***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
>
@MAnalytics is the same account as @geoMADE - the account name was changed some time ago.
>
***## ------------------------------------------***
***2. Reviewer's comment:***
***## ------------------------------------------***
>
> ### Functionality
> * [x]  **Installation:** Does installation proceed as outlined in the documentation?
> * [ ]  **Functionality:** Have the functional claims of the software been confirmed?
>
> One common theme I've seen in this package is that there is a lot of code being printed out to the console when functions are being run, and there is no way to silence this. For example:
>
> ```
> example_clust <- akmedoids.clust(all_traj, k=c(3,10), crit = "Silhouette")
> ```
>
> returns
>
> ```
> [1] "solution of k = 3 determined!"
> [1] "solution of k = 4 determined!"
> [1] "solution of k = 5 determined!"
> [1] "solution of k = 6 determined!"
> [1] "solution of k = 7 determined!"
> [1] "solution of k = 8 determined!"
> [1] "solution of k = 9 determined!"
> [1] "solution of k = 10 determined!"
> [1] "Suggested optimal solution contains 3 clusters. See the plot for further examination!"
> ```
>
> These print messages are useful, but there should be a `verbose` or `quiet` argument to suppress these. This information should be also stored in the output of the object.
>
> This function also opened a graphic window and plot. There is no way for the user to suppress this. The output should be stored in the object, and if you really want to show the plot to uses, have an option to print the plot when running the function, perhaps with an argument like `show_plot = TRUE`.
>
> Reading the documentation of the package, the function returns different output based on the value of K. The documentation reads:
>
> > #' @return If \code{k} is a vector of two numbers (see param.
> > #' \code{k} details above), the output is a graphical plot of
> > #' the quality scores of the cluster solutions. If \code{k} is
> > #' an exact integer number of clusters, the function returns
> > #' trajectory labels indicating the group membership of the
> > #' corresponding trajectory in the \code{traj} object.
>
> I think that this makes the software harder to use, as there is an implied workflow that could be made simpler and easier to use.
>
> It seems then that the next step is to run the same code again but with `k` = 3. I am not sure why this needs to be rerun, shouldn't this be stored as the output? For example, a vector of classes for each value of `k`.
>
> Running the function again with `k = 3`, I get the following:
>
> ```
> example_clust_3 <- akmedoids.clust(all_traj, k=c(3), crit = "Silhouette")
> example_clust_3
> ```
>
> Finally to summarise the information we run:
>
> ```
> #get the optimal cluster solution
> clustr <- akmedoids.clust(all_traj, k=3)
> clustr <- as.vector(clustr$memberships)$alphabetic_Labels
>
> ##get group statistics
> descrStats <- statPrint(clustr, all_traj,
>                        id_field = FALSE, showplots=FALSE)
>
> print(descrStats)
> ```
>
> I believe that this can be improved in a few ways.
>
> Firstly, writing
>
> ```r
> clustr <- akmedoids.clust(all_traj, k=3)
> clustr <- as.vector(clustr$memberships)$alphabetic_Labels
> ```
>
> should be avoided as you are overwriting the object.
>
> Secondly, the user should not need to perform special data manipulation (such as accessing a nested list with (`as.vector(clustr$memberships)$alphabetic_Labels`) in order to summarise the data.
>
> This could be improved by writing a special `print` method for the software. This would mean that printing `clustr` out to the console, e.g.,
>
> ```r
> clustr <- akmedoids.clust(all_traj, k=3)
> clustr
> ```
>
> Outputting the `clustr` object in the console, it could return the information that `statPrint` returns. You can read more about creating custom `print` methods here - https://adv-r.hadley.nz/s3.html, https://blog.earo.me/2019/11/03/practical-guide-to-s3/.
>
> It seems that there is a workflow implied:
>
> 1. Test multiple different values of `k` with `akmedoids.clust`
>
> * explore the plots of the optimal solution based on a certain criterion.
> * decide on ideal value of K
>
> 1. Run clustering for the optimally identified value of `k`
>
> * using `akmedoids.clust` again
>
> 1. summarise this information using `statPrint`
>
> I think that this could be improved in the following way.
>
> 1. Provide a function for exploring different values of K
> 2. Provide a function for identifying the cluster labels
> 3. Provide good print methods, so the output that is given to the console could be delivered via the print method.
>
> I suggestion is the output of this function should contain possible memberships for
>
> I also note that `statPrint` is noted as performing two tasks - (1) generating statistics of groups, and (2), creating plots.
>
> In my opinion the output of `statPrint` and `akmedoids.clust` functions should be separated into two functions, or at least provide options for the plots to be turned off. If they are turned into functions, I recommend:
>
> * One function for identifying clusters or summary statistics that return a data frame output or a list.
> * And one function that focuses on creating graphics.
>
***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
>
Thank you very much for taking the time to explore these functions and suggesting ways to improve their functionalities.

First, in order to improve the usability of the package, we make the following changes to the `akmedoids.clust` and `statPrint` functions, and how they can be deployed.

1. ***Change of names***: We changed the names of `akmedoids.clust` and `statPrint` functions to more intuitive names: `akClustr` and `print_akStats` (**a special print function**) functions, respectively.

2. ***Workflow***:  The above two functions can now be deployed to perform clustering and visualize the results, respectively, following ***two simple steps*** as follows:

(***a***) Use `akClustr` function to generate the solutions for different values of `k`. For example,

 ```
> aksolution <- akClustr(trajectry, k = c(3,7), verbose = FALSE, quality_Plot=FALSE)
```
To print the results out to the console, a user can just type:

```
> aksolution
```
(This will print the solutiond for all values of `k` and also indicates which amongst them is the optimal solution – i.e. `$optimal_k`).

As seen above, the function now includes `verbose` argument to suppress the messages during processing, and a `quality_Plot` argument to control the plot of quality criteria. The output of `akClustr` function is exported as object of class `akObject`.

(***b***) Given an object of class `akObject` (as in the above), use `print_akStats` function to generate the properties (i.e. descriptive and change statistics) of any specified `k`-cluster solution (i.e. optimal or not). In the example above, the properties of the `k =4` solution for example can be printed as:

```
> print_akStats(aksolution, k = 4, type="lines", y.scaling="fixed", show_Plots = TRUE)
```
Note that the function now includes `show_Plots` argument for controlling the clusters plots.

***## ------------------------------------------***
***3. Reviewer's comment:***
***## ------------------------------------------***

> I have some other comments on the functions in this package:
>
> * `akmedoids::alphaLabel()`
>
>   * I'm not sure if the output of this needs to be a list, it could instead just be the vector of characters.
>   * Is it possible to provide an example of the use of this function?

***## ------------------------------------------***
***Response:***
***## ------------------------------------------***

The output is actually a vector of characters.

The goal of `akmedoids::alphaLabel()` function is to allow the transformation of any numeric cluster labels (particularly of external packages, such as `kml`) into alphabetic labels. The goal is to allow a comparison of results from such packages (e.g. `kml`) with the results from our own package. We have now included an @example usage of this type of scenario in the `alphaLabel()` function, by showing how the numeric cluster labels of `kml::affectIndiv()` function can be transformed into alphabetic labels. We deployed a real-life `Time-at-risk for the adjudicated Toronto Youth` datasets (`Nielsen, J. 2018`).

The `alphaLabel()` function was originally used within our `akClustr` function (see line 331 of `akClustr` function).

***## ------------------------------------------***
***4. Reviewer's comment:***
***## ------------------------------------------***
> * `akmedoids::dataImputation()`
>
>   * see note in "common issues" about printing messages
>
***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
>
A `verbose` argument has now been added to the `akmedoids::dataImputation()` function to control the output messages.

***## ------------------------------------------***
***5. Reviewer's comment:***
***## ------------------------------------------***

> * `akmedoids::elbowPoint()`
>
>   * In example documentation, the function is applied to the plot (which is great!) but is not demonstrated in other contexts. I suggest providing an example outside of plotting as well.
>
***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
>
We have now modified the @example of the `elbowPoint()` function by using a real-life data set (i.e. `Time-at-risk for the adjudicated Toronto Youth` datasets). We deploy `akClustr` function to demonstrate how the `elbowPoint()` function helps the `Silhouette` criterion to determine the optimality of `k` values.
>
***## ------------------------------------------***
***6. Reviewer's comment:***
***## ------------------------------------------***
>
> `akmedoids::outlierDetect()`
>
> * This prints out several messages and objects, see below note on messages.

***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
A `verbose` argument is added to control the print out messages.

***## ------------------------------------------***
***7. Reviewer's comment:***
***## ------------------------------------------***
> * `akmedoids::props()`
>
>   * states output is a matrix, but is a data.frame
>
***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
>
Thanks. Corrected accordingly.
>
***## ------------------------------------------***
***8. Reviewer's comment:***
***## ------------------------------------------***
>
> * `akmedoids::rates()`
>
>   * This function doesn't return a matrix, instead returning a vector of common IDs. I am not sure what these mean, as their behaviour is not documented in the helpfile.
>
***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
>
In our checks the function seems to be working fine. Please kindly see the @example provided under the function.
>
***## ------------------------------------------***
***9. Reviewer's comment:***
***## ------------------------------------------***
>
> * `akmedoids::removeRowsN()`
>
>   * Could the documentation clarify how is this different to `na.omit()`?
>
***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
>
We have now added the following statement in the description of the `removeRowsN()` function:

*"`... . The function is also able to remove records with `Inf` entries, distinguishing it from the popular `na.omit()` function in` R".
>
***## ------------------------------------------***
***10. Reviewer's comment:***
***## ------------------------------------------***
>
> * `akmedoids::wSpaces()`
>
>   * The example for this function doesn't contain whitespace characters
>   * Could the documentation explain how this is different to `trimws` function in R?
>
***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
We have added a new sample data `traj_wSpaces.rda` which contains whitespaces in order to demonstrate the functionality of this function. Please, see the @example. Note: the `wSpaces()` function is found to perform the same task as the `trimws()` function.
>
***## ------------------------------------------***
***11. Reviewer's comment:***
***## ------------------------------------------***

> * `akmedoids::statPrint()`
>
>   * It is very neat that this function can generate a nice looking ggplot, however it would be great if that plot was instead a separate function
>
***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
Thanks for the comment. Please, see above. Please, also note that `akmedoids::statPrint()` has now been changed to `akmedoids::print_akStats()`
>
***## ------------------------------------------***
***12. Reviewer's comment:***
***## ------------------------------------------***
>
> #### Common issues
> * plots all open new graphics devices rather than using standard `print` methods, which can be problematic when trying to include them in documents
> * no way to silence console output - suggest adding a `quiet` or `verbose` option, and to use `message` when printing messages to the console.
>
>   * e.g., the `rates`, `akmedoids.clust` functions
>
***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
The ***Common issues*** have all been addressed (please, see above)
>
> * [x]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> > Yes.
>
> ### Documentation
> * [ ]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
>
> The statement in the README says:
>
> > identify clusters of trajectories with similar long-term linear trends over time, providing an improved cluster identification as compared with the classic kmeans algorithm
>
> * [x]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
>
> Yes
>
***## ------------------------------------------***
***13. Reviewer's comment:***
***## ------------------------------------------***

> * [ ]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>
> > Although the author provides instructions in the README, but these contain no output of the function. This should be updated to make the README a `README.Rmd` function. See `usethis::use_readme_rmd()` as an approach on doing this.
>
> > In addition, the README simulates some example data for use - this should be encapsulated as an example dataset in the package that can be loaded.
>
***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
>
Thank you for the suggestion.

We have now provided a `README.Rmd`, addressing the concerns above. We also update the package vignette accordingly. Specifically, we detailed the following in the `README.Rmd` document:

 * Installation instructions
* Sample codes for simulating example trajectory groups,
* Codes to visualize simulated trajectory groups.
* Code to cluster the groups, print cluster properties, and plot the clusters.

***## ------------------------------------------***
***14. Reviewer's comment:***
***## ------------------------------------------***

> * [ ]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> * example documentation has unnecessary `traj <- traj` line in "examples"
> * The datasets are not described in the helpfiles, and state that they are matrices when they are data frames. This data should be documented more thoroughly (see https://r-pkgs.org/data.html#documenting-data for details on how to document data, and https://github.com/jforbes14/eechidna/blob/master/R/data.R for an example documentation of data).
>
***## ------------------------------------------***
***Response:***
***## ------------------------------------------***
>
* Unnecessary lines have been removed.
* We have now followed the example in https://github.com/jforbes14/eechidna/blob/master/R/data.R and provide detailed documentation of all the sample data sets included in the package. Please, see `./R/data.R`

***Other additional revision to note:***
* We renamed `population.rda` as `popl.rda` as there is a clash with the `population.rda` in one of the dependency packages.

> * [x]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?
>
> > There are tests
>
[…]
