### 2020-11-17 #2379 @njtierney
# review response

Thank you for your response! I'm looking forward to using this in my own work.

I greatly appreciate your effort in addressing my points, for readability, I will address the relevant points freshly below:

## Function arguments should have consistent names

- This package uses camelCase for it's functions names, (akClustr, alphaLabel, etc)
- Some of the function arguments use mixed names, such as the function, `print_akStats` has the arguments, "akObject", "k", "reference", "N.quant",, "show_Plots", "type", "y.scaling". These arguments should be either
  - all camelCase (akObject, yScaling, nQuant), or
  - all lower case separated by an underscore (e.g., ak_object, y_scaling)
- It should not be a mix of changing case and separator, as this is hard for the user to predict (e.g., N.quant, show_Plots)

## Print methods

This isn't required to be addressed for the review, but it might be helpful for your own work - your function, `print_akStats` can be written in a way so that when you print an object created by `akClustr`, it will do the nice printing when output into the console. Again, not needed for this review, but worth knowing about.

## Separate functions for plotting

- The plot methods currently open up a new device when plots are shown, from this code:

```
  if(show_Plots==TRUE){
    #plot
    flush.console()
    dev.new(width=3, height=3)
    print(plt)
  }
```

This code could instead be:

```
  if(show_Plots==TRUE){
    print(plt)
  }
```

as there isn't a need to flush the console or create a specific device

I also recommend that this plotting function is separated, perhaps as a `plot_akStats` function. This can then be called inside your print method if you prefer, but having just the function to plot the information would be really nice.

## regarding `akmedoids::rates()`

I stated in my previous review:

>   * This function doesn't return a matrix, instead returning a vector of common IDs. I am not sure what these mean, as their behaviour is not documented in the helpfile.

Your response:

> In our checks the function seems to be working fine. Please kindly see the @example provided under the function.

When I run the example code, I get this output:

```
$common_ids
[1] "E01012628" "E01004768" "E01004803" "E01004804" "E01004807" "E01004808"
[7] "E01004788" "E01004790" "E01004805"

$noncommon_ids_from_trajectories
[1] "E01004806"

$noncommon_ids_from_denominator
[1] "E01004809" "E01004899"

$rates_data
  location_ids X2001 X2002 X2003 X2004 X2005 X2006 X2007 X2008 X2009
1    E01012628  3.60  2.86  1.00  2.29  4.67  4.00  1.00  4.00  2.00
2    E01004768  4.00  6.40  3.63  2.29  2.00  3.56  0.80  2.18  0.00
3    E01004803  2.86  2.48  4.53  3.35  3.75  3.30  3.29  4.34  2.00
4    E01004804  2.29  1.60  0.00  4.71  0.89  1.26  2.40  2.29  2.91
5    E01004807  5.09  1.92  5.14  1.55  0.94  1.69  2.40  1.12  0.70
6    E01004808  3.76  2.42  5.00  3.61  4.80  2.21  3.09  3.56  1.85
7    E01004788  0.67  1.36  1.82  1.90  2.40  1.84  1.33  2.35  2.00
8    E01004790  6.40  3.64  5.33  2.46  0.57  2.67  3.00  0.47  0.44
9    E01004805  6.00  0.00  2.00  4.00  2.00  0.00  2.00  8.00  0.00
```

In the documentation, can you describe what, `$common_ids` `$noncommon_ids_from_trajectories`, and `$noncommon_ids_from_denominator` are?

## Regarding the README.Rmd

- A very minor thing - the README.Rmd file uses html_document instead of github_document (commented out [here](https://github.com/MAnalytics/akmedoids/blob/master/README.Rmd#L2)) - this should then generate the R output that can be seen in the README.md
- Not necessary for this review at all, but the authors might also be interested in looking at the `pkgdown` R package for automatically generating websites for R packages. See https://pkgdown.r-lib.org/ for more details.
