+++ man/akClustr.Rd
@@ -1,11 +1,11 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/akClustr.R
-\name{akClustr}
-\alias{akClustr}
+% Please edit documentation in R/akclustr.R
+\name{akclustr}
+\alias{akclustr}
 \title{Anchored k-medoids clustering}
 \usage{
-akClustr(traj, id_field = FALSE, method = "linear",
-k = c(3,6), crit="Silhouette", verbose = TRUE, quality_Plot=FALSE)
+akclustr(traj, id_field = FALSE, method = "linear",
+k = c(3,6), crit="Silhouette", verbose = TRUE, quality_plot=FALSE)
 }
 \arguments{
 \item{traj}{[matrix (numeric)]: longitudinal data. Each row represents
@@ -34,7 +34,7 @@ to use for assessing the quality of the cluster solutions, when
 \code{crit="Silhouette"}, use the average Silhouette width
 (\code{Rousseeuw P. J. 1987}). Using the \code{"Silhouette"} criterion,
 the optimal value of \code{k} can be determined as the elbow point of
-the curve. Other valid criterion is the "Caliński_Harabasz"
+the curve. Other valid criterion is the "Calinski_Harabasz"
 (\code{Caliński T. & Harabasz J. 1974}) in which the maximum score
 represents the point of optimality. Having determined the optimal
 \code{k}, the function can then be re-run, using the exact (optimal)
@@ -43,11 +43,11 @@ value of \code{k}.}
 \item{verbose}{to suppress output messages (to the console)
 during clustering. Default: \code{TRUE}.}
 
-\item{quality_Plot}{Whether to show plot of quality criteria across
+\item{quality_plot}{Whether to show plot of quality criteria across
 different values of \code{k}. Default: \code{FALSE}.}
 }
 \value{
-generates an \code{akObject} consisting of the
+generates an \code{akobject} consisting of the
 cluster solutions at the specified values of \code{k}. Also,
 the graphical plot of the quality scores of the cluster
 solutions.
@@ -73,16 +73,16 @@ the selection of observations (medoids) as oppose to centroids.
 
 data(traj)
 
-trajectry <- dataImputation(traj, id_field = TRUE, method = 2,
+trajectry <- data_imputation(traj, id_field = TRUE, method = 2,
 replace_with = 1, fill_zeros = FALSE)
 
 trajectry <- props(trajectry$CompleteData, id_field = TRUE)
 
 print(trajectry)
 
-output <- akClustr(trajectry, id_field = TRUE,
+output <- akclustr(trajectry, id_field = TRUE,
 method = "linear", k = c(3,7), crit='Calinski_Harabasz',
-verbose = FALSE, quality_Plot=FALSE)
+verbose = FALSE, quality_plot=FALSE)
 
 print(output)
 
