+++ man/traj_w_spaces.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/data.R
 \docType{data}
-\name{traj_wSpaces}
-\alias{traj_wSpaces}
+\name{traj_w_spaces}
+\alias{traj_w_spaces}
 \title{Sample longitudinal dataset containing whitespaces}
 \format{
 A dataframe with the following variables:
@@ -21,7 +21,7 @@ A dataframe with the following variables:
     }
 }
 \usage{
-traj_wSpaces
+traj_w_spaces
 }
 \description{
 Longitudinal dataset with both trailing and leading
