+++ man/props.Rd
@@ -37,7 +37,7 @@ by the sum of the column where it is located, i.e.
 }
 \examples{
 
-trajectry <- dataImputation(traj, id_field = TRUE, method = 2,
+trajectry <- data_imputation(traj, id_field = TRUE, method = 2,
 replace_with = 1, fill_zeros = FALSE) #filling the missing values
 
 trajectry <- props(trajectry$CompleteData, id_field = TRUE,
