+++ man/plot_akstats.Rd
@@ -1,10 +1,17 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/print_akstats.R
-\name{print_akstats}
-\alias{print_akstats}
-\title{Descriptive (Change) statistics}
+% Please edit documentation in R/plot_akstats.R
+\name{plot_akstats}
+\alias{plot_akstats}
+\title{Plot of cluster groups.}
 \usage{
-print_akstats(ak_object, k = 3, reference = 1, n_quant = 4, show_plots = FALSE)
+plot_akstats(
+  ak_object,
+  k = 3,
+  reference = 1,
+  n_quant = 4,
+  type = "lines",
+  y_scaling = "fixed"
+)
 }
 \arguments{
 \item{ak_object}{An output of \code{\link{akclustr}} function.
@@ -31,28 +38,30 @@ This result can be used to further categorize the groups into
 'classes'. For example, groups that fall within the \code{1st}
 quartile may be classified as 'Stable' groups (Adepeju et al. 2020).}
 
-\item{show_plots}{[TRUE or FALSE] To display cluster plots.
-Line plots are displayed by default. Please, see
-\code{plot_akstats} function for more plot options.
-Defaults \code{FALSE}}
+\item{type}{[character] plot type. Available options are:
+\code{"lines"} and \code{"stacked"}.}
+
+\item{y_scaling}{[character] works only if \code{type="lines"}.
+\code{y_scaling} set the vertical scales of the cluster panels.
+Options are: \code{"fixed"}: uses uniform scale for all panels,
+\code{"free"}: uses variable scales for panels.}
 }
 \value{
 A plot showing group membership or sizes (proportion)
 and statistics.
 }
 \description{
-This function perform two tasks:
-(i) it generate the descriptive and change statistics
-of groups, particularly suited for the outputs form
-the \code{\link{akclustr}} function, and
-(ii) generates the plots of the groups (performances).
+Takes the 'ak_object' from the
+\code{'akclustr'} as input and produce either the 'line' plot
+or 'stacked' histogram.
 }
 \details{
-Generates the plot of trajectory groupings.
-Given an \code{ak_object} class (from
-the \code{akclustr} function), this function show the
-plots of cluster groups. The plot component draws from
-\code{plot_akstats} function.
+Generates the plots of cluster groups - same plots
+generated by the \code{'show_plots'} argument of \code{print_akstats}.
+The function draw from the functionalities of the
+\code{ggplot2} library.
+For a more customized visualisation, we recommend that users
+deploy \code{ggplot2} directly (\code{Wickham H. (2016)}).
 }
 \examples{
 
@@ -66,9 +75,13 @@ print(trajectry$CompleteData)
 trajectry <- props(trajectry$CompleteData, id_field = TRUE)
 
 aksolution <- akclustr(trajectry, id_field = TRUE,
-    method = "linear", k = c(3,5), crit='Calinski_Harabasz')
+method = "linear", k = c(3,5), crit='Calinski_Harabasz')
 
-print_akstats(aksolution, k = 4, show_plots=FALSE)
+plot_akstats(aksolution, k = 4, type="lines",
+y_scaling="fixed")
+
+plot_akstats(aksolution, k = 4, reference = 1,
+n_quant = 4, type="stacked")
 
 }
 \references{
