+++ man/remove_rows_n.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/removeRowsN.R
-\name{removeRowsN}
-\alias{removeRowsN}
+% Please edit documentation in R/remove_rows_n.R
+\name{remove_rows_n}
+\alias{remove_rows_n}
 \title{Removes rows that contain 'NA' and/or 'Inf' entries}
 \usage{
-removeRowsN(traj, id_field=TRUE, remove=1)
+remove_rows_n(traj, id_field=TRUE, remove=1)
 }
 \arguments{
 \item{traj}{[data.frame (numeric)]: longitudinal data.
@@ -38,5 +38,5 @@ complete observations.
 
 data(traj)
 
-removeRowsN(traj, id_field=TRUE, remove=3)
+remove_rows_n(traj, id_field=TRUE, remove=3)
 }
