+++ tests/testthat/test_elbow_point.R
@@ -1,11 +1,11 @@
-context("Testing elbowPoint.R function")
+context("Testing elbow_point.R function")
 
 x <- runif(100, min=-2, max=3)
 y <- -exp(-x) * (1+rnorm(100)/3)
 plot(x, y)
 
 # get the elbow points
-elbP <- elbowPoint(x,y)
+elbP <- elbow_point(x,y)
 
 #sign of most extreme derivative
 pick.sign <- function(x) {
