+++ README.Rmd
@@ -146,11 +146,11 @@ ggplot(simulated_long, aes(x = Time, y = value, group=ID, color=Groups)) +
 
 ### Performing clustering using `akmedoids`
 
-The `akmedoids::akClustr()` is used to perform the clustering analysis. Using the simulated data, we write:
+The `akmedoids::akclustr()` is used to perform the clustering analysis. Using the simulated data, we write:
 
 ```{r, echo=TRUE, message=TRUE, eval=TRUE}
 
-output <- akClustr(Import_simulated, id_field=TRUE, verbose = FALSE, k=c(3,12), crit = "Silhouette",
+output <- akclustr(Import_simulated, id_field=TRUE, verbose = FALSE, k=c(3,12), crit = "Silhouette",
                   quality_Plot=FALSE)
 
 #RemarkS: The results can be printed out to the console by typing `output`
@@ -179,17 +179,17 @@ names(output)
 
   + `qualityCrit.List` - the estimated quality of cluster solutions by `k` values.
 
-  + `QltyPlot` - the plot of `qualityCrit.List`, with a red vertical line to indicate the optimal value of `k`.
+  + `qltyplot` - the plot of `qualityCrit.List`, with a red vertical line to indicate the optimal value of `k`.
 
 
 *Accessing the optimal solution*
 
-The plot of `qualityCrit.List` can be displayed by setting the `quality_Plot` argument as `TRUE`. Otherwise, a user may access the plot by printing the variable `output$QltyPlot` to the console. For example,
+The plot of `qualityCrit.List` can be displayed by setting the `quality_plot` argument as `TRUE`. Otherwise, a user may access the plot by printing the variable `output$qltyplot` to the console. For example,
 
 
 ```{r, fig.width=7, fig.height=5, fig.fullwidth=TRUE}
 
-output$QltyPlot
+output$qltyplot
 
 #the intersection of the red line with the x-axis indicates the optimal k value.
 ```
@@ -199,13 +199,13 @@ The optimal solution is determined to be 3, as also stored by the variable `outp
 
 ### Printing Cluster Stats.
 
-The properties (i.e. the descriptive and change statistics) of any of the cluster solutions (i.e. for any value of `k`) can be accessed using the special `print` function `print_akStats`. The print function takes the `akObject` class generated by the `akmedoids::akClustr` function as input as follows: 
+The properties (i.e. the descriptive and change statistics) of any of the cluster solutions (i.e. for any value of `k`) can be accessed using the special `print` function `print_akstats`. The print function takes the `akobject` class generated by the `akmedoids::akclustr` function as input as follows: 
 
 ```{r, echo=TRUE, message=FALSE, eval=TRUE}
 
 #Specifying the optimal solution, output$optimal_k (i.e. `k = 3`)
 
-prpties = print_akStats(output, k = 3, type="lines", y.scaling="fixed", show_Plots = FALSE)
+prpties = print_akstats(output, k = 3, show_plots = FALSE)
 
 ```
 Accessing the variables of `prpties`:
@@ -218,11 +218,10 @@ names(prpties)
 
 * output Variables described as follow:
 
-  + `descriptiveStats` - descriptive statistics of the cluster.
+  + `descriptive_stats` - descriptive statistics of the cluster.
   
-  + `changeStats` - change statistics of the cluster
-  
-  + `cluster_Plot` - the line plot or stacked histogram of cluster groups.
+  + `change_stats` - change statistics of the cluster
+
 
 ### Printing the group properties
 
@@ -230,7 +229,7 @@ names(prpties)
 
[…]
