+++ NAMESPACE
@@ -1,16 +1,18 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(print_akStats,default)
-export(akClustr)
-export(alphaLabel)
-export(dataImputation)
-export(elbowPoint)
-export(outlierDetect)
-export(print_akStats)
+S3method(plot_akstats,default)
+S3method(print_akstats,default)
+export(akclustr)
+export(alpha_label)
+export(data_imputation)
+export(elbow_point)
+export(outlier_detect)
+export(plot_akstats)
+export(print_akstats)
 export(props)
 export(rates)
-export(removeRowsN)
-export(wSpaces)
+export(remove_rows_n)
+export(w_spaces)
 importFrom(Hmisc,cut2)
 importFrom(clusterCrit,intCriteria)
 importFrom(ggplot2,aes)
