+++ R/akClustr.R
@@ -27,17 +27,17 @@
 #' \code{crit="Silhouette"}, use the average Silhouette width
 #' (\code{Rousseeuw P. J. 1987}). Using the \code{"Silhouette"} criterion,
 #' the optimal value of \code{k} can be determined as the elbow point of
-#' the curve. Other valid criterion is the "Caliński_Harabasz"
+#' the curve. Other valid criterion is the "Calinski_Harabasz"
 #' (\code{Caliński T. & Harabasz J. 1974}) in which the maximum score
 #' represents the point of optimality. Having determined the optimal
 #' \code{k}, the function can then be re-run, using the exact (optimal)
 #' value of \code{k}.
 #' @param verbose to suppress output messages (to the console)
 #' during clustering. Default: \code{TRUE}.
-#' @param quality_Plot Whether to show plot of quality criteria across
+#' @param quality_plot Whether to show plot of quality criteria across
 #' different values of \code{k}. Default: \code{FALSE}.
-#' @usage akClustr(traj, id_field = FALSE, method = "linear",
-#' k = c(3,6), crit="Silhouette", verbose = TRUE, quality_Plot=FALSE)
+#' @usage akclustr(traj, id_field = FALSE, method = "linear",
+#' k = c(3,6), crit="Silhouette", verbose = TRUE, quality_plot=FALSE)
 #' @details This function works by first approximating the trajectories
 #' based on the chosen parametric forms (e.g. linear), and then partitions
 #' the original trajectories based on the form groupings, in similar
@@ -50,20 +50,20 @@
 #'
 #' data(traj)
 #'
-#' trajectry <- dataImputation(traj, id_field = TRUE, method = 2,
+#' trajectry <- data_imputation(traj, id_field = TRUE, method = 2,
 #' replace_with = 1, fill_zeros = FALSE)
 #'
 #' trajectry <- props(trajectry$CompleteData, id_field = TRUE)
 #'
 #' print(trajectry)
 #'
-#' output <- akClustr(trajectry, id_field = TRUE,
+#' output <- akclustr(trajectry, id_field = TRUE,
 #' method = "linear", k = c(3,7), crit='Calinski_Harabasz',
-#' verbose = FALSE, quality_Plot=FALSE)
+#' verbose = FALSE, quality_plot=FALSE)
 #'
 #' print(output)
 #'
-#' @return generates an \code{akObject} consisting of the
+#' @return generates an \code{akobject} consisting of the
 #' cluster solutions at the specified values of \code{k}. Also,
 #' the graphical plot of the quality scores of the cluster
 #' solutions.
@@ -86,9 +86,9 @@
 #' @export
 
 
-akClustr <- function(traj, id_field = FALSE, method = "linear",
+akclustr <- function(traj, id_field = FALSE, method = "linear",
                             k = c(3,6), crit = "Silhouette",
-                            verbose = TRUE, quality_Plot=FALSE){
+                            verbose = TRUE, quality_plot=FALSE){
 
   qualityCrit <- 0
 
@@ -106,11 +106,11 @@ akClustr <- function(traj, id_field = FALSE, method = "linear",
     #list k
     k_ <- k[1]:k[2]
 
-    #to check that quality_Plot' is used correctly.
-    if(k[1]==k[2] & quality_Plot == TRUE){
-      stop(paste("...'quality_Plot==TRUE' argument not applicable.",
+    #to check that quality_plot' is used correctly.
+    if(k[1]==k[2] & quality_plot == TRUE){
+      stop(paste("...'quality_plot==TRUE' argument not applicable.",
                  "Provide a set of k values (See documentation of",
-                 "'akClustr' function).", sep=" "))}
+                 "'akclustr' function).", sep=" "))}
 
     #check if unacceptable value of k in inputted
 
[…]
