### 2020-11-18 e5e8932
delete alp lab
