+++ R/print_akStats.R
@@ -3,26 +3,20 @@
 #' @description This function perform two tasks:
 #' (i) it generate the descriptive and change statistics
 #' of groups, particularly suited for the outputs form
-#' the \code{\link{akClustr}} function, and
+#' the \code{\link{akclustr}} function, and
 #' (ii) generates the plots of the groups (performances).
-#' @param akObject An output of \code{\link{akClustr}} function.
+#' @param ak_object An output of \code{\link{akclustr}} function.
 #' The object contains individual trajectories and their cluster
 #' solution(s) at the specified values of \code{k}. Also, includes
 #' the optimal value of \code{k} based on the criterion specified.
 #' at (different) values of \code{k} the \code{traj}.
 #' @param k [integer] \code{k} cluster to generate its solution.
-#' @param type [character] plot type. Available options are:
-#' \code{"lines"} and \code{"stacked"}.
-#' @param y.scaling [character] works only if \code{type="lines"}.
-#' \code{y.scaling} set the vertical scales of the cluster panels.
-#' Options are: \code{"fixed"}: uses uniform scale for all panels,
-#' \code{"free"}: uses variable scales for panels.
 #' @param reference [numeric] Specifying the reference line from
 #' which the direction of each group is measured. Options are:
 #' \code{1}: slope of mean trajectory, \code{2}: slope of medoid
 #' trajectory, \code{3}: slope of a horizontal line
 #' (i.e. slope = 0). Default: \code{1}.
-#' @param N.quant [numeric] Number of equal intervals (quantiles)
+#' @param n_quant [numeric] Number of equal intervals (quantiles)
 #' to create between the reference line \code{(R)} and the medoids
 #' \code{(M)} of the most-diverging groups of both sides of
 #' \code{(R)}. Default is \code{4} - meaning quartile subdivisions
@@ -30,39 +24,33 @@
 #' returns the quartile in which the medoid of each group falls.
 #' This result can be used to further categorize the groups into
 #' 'classes'. For example, groups that fall within the \code{1st}
-#' quartile may be classified as 'Stable' groups (Adepeju et al. 2019).
-#' @param show_Plots [TRUE or FALSE] To display cluster plots.
-#' Defaults \code{TRUE}
-#' @s3method print_akStats
+#' quartile may be classified as 'Stable' groups (Adepeju et al. 2020).
+#' @param show_plots [TRUE or FALSE] To display cluster plots.
+#' Line plots are displayed by default. Please, see
+#' \code{plot_akstats} function for more plot options.
+#' Defaults \code{FALSE}
+#' @s3method print_akstats
 #' @examples
 #'
 #' data(traj)
 #'
-#' trajectry <- dataImputation(traj, id_field = TRUE, method = 1,
+#' trajectry <- data_imputation(traj, id_field = TRUE, method = 1,
 #' replace_with = 1, fill_zeros = FALSE)
 #'
 #' print(trajectry$CompleteData)
 #'
 #' trajectry <- props(trajectry$CompleteData, id_field = TRUE)
 #'
-#' akSolution <- akClustr(trajectry, id_field = TRUE,
-#' method = "linear", k = c(3,5), crit='Calinski_Harabasz')
+#' aksolution <- akclustr(trajectry, id_field = TRUE,
+#'     method = "linear", k = c(3,5), crit='Calinski_Harabasz')
 #'
-#' print_akStats(akSolution, k = 4, type="lines",
-#' y.scaling="fixed", show_Plots = TRUE)
+#' print_akstats(aksolution, k = 4, show_plots=FALSE)
 #'
-#' print_akStats(akSolution, k = 4, reference = 1,
-#' show_Plots = TRUE, N.quant = 4, type="stacked")
-#'
[…]
