+++ R/props.R
@@ -14,7 +14,7 @@
 #' @usage props(traj, id_field = TRUE, scale = 1, digits = 4)
 #' @examples
 #'
-#' trajectry <- dataImputation(traj, id_field = TRUE, method = 2,
+#' trajectry <- data_imputation(traj, id_field = TRUE, method = 2,
 #' replace_with = 1, fill_zeros = FALSE) #filling the missing values
 #'
 #' trajectry <- props(trajectry$CompleteData, id_field = TRUE,
