### 2020-11-18 e856d48
create remove_rows
