+++ README.md
@@ -1,29 +1,67 @@
-## 'akmedoids' package
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+# akmedoids
+
+<!-- badges: start -->
+
+<!-- badges: end -->
 
 An R package for analyzing and clustering longitudinal data
 
 ### Description
 
-The `akmedoids` package advances the clustering of longitudinal datasets in order to identify clusters of trajectories with similar long-term linear trends over time, providing an improved cluster identification as compared with the classic kmeans algorithm. The package also includes a set of functions for addressing common data issues, such as missing entries and outliers, prior to conducting advance longitudinal data analysis. One of the key objectives of this package is to facilitate easy replication of a recent paper which examined small area inequality in the crime drop (Adepeju et al. 2020). Many of the functions provided in the `akmedoids` package may be applied to longitudinal data in general. 
+The `akmedoids` package advances the clustering of longitudinal datasets
+in order to identify clusters of trajectories with similar long-term
+linear trends over time, providing an improved cluster identification as
+compared with the classic kmeans algorithm. The package also includes a
+set of functions for addressing common data issues, such as missing
+entries and outliers, prior to conducting advance longitudinal data
+analysis. One of the key objectives of this package is to facilitate
+easy replication of a recent paper which examined small area inequality
+in the crime drop (Adepeju et al. 2020). Many of the functions provided
+in the `akmedoids` package may be applied to longitudinal data in
+general.
 
-**For more information and usability, check out details on [CRAN](https://cran.r-project.org/web/packages/akmedoids/index.html).**
+**For more information and usability, check out details on
+[CRAN](https://cran.r-project.org/web/packages/akmedoids/index.html).**
 
-### Installation 
+### Installation from `CRAN`
 
-```R
-#installation from `CRAN`
-#From an R console, type:
-install.packages("akmedoids")
+From an R console, type:
+
+``` r
+#install.packages("akmedoids")
 library(akmedoids)
+
+#Other libraries
+library(tidyr)
+library(ggplot2)
+#> Warning: package 'ggplot2' was built under R version 4.0.3
+library(reshape)
+#> Warning: package 'reshape' was built under R version 4.0.3
+library(readr)
 ```
-To install the development version of the package, type `remotes::install_github("MAnalytics/akmedoids")`. Please, report any installation problems in the issues.
+
+To install the development version of the package, type
+`remotes::install_github("MAnalytics/akmedoids")`. Please, report any
+installation problems in the issues.
 
 ### Example usage:
 
-Given a longitudinal datasets, the following is an example of how `akmedoids` could be used to extract clusters of trajectories with similar long-term trends over time. 
+Given a longitudinal datasets, the following is an example of how
+`akmedoids` could be used to extract clusters of trajectories with
+similar long-term trends over time. We will use a simulated dataset
+(named `simulated.rda`) stored in the `data/` directory.
 
-```R
-#simulate some datasets
+### Generating artificial dataset
+
[…]
