### 2020-11-18 524632a
update manuals
