+++ man/data_imputation.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/dataImputation.R
-\name{dataImputation}
-\alias{dataImputation}
+% Please edit documentation in R/data_imputation.R
+\name{data_imputation}
+\alias{data_imputation}
 \title{Data imputation for longitudinal data}
 \usage{
-dataImputation(traj, id_field = FALSE, method = 2,
+data_imputation(traj, id_field = FALSE, method = 2,
 replace_with = 1, fill_zeros = FALSE, verbose=TRUE)
 }
 \arguments{
@@ -71,7 +71,7 @@ estimate the missing values.
 
 # Using the example 'traj' datasets
 
-imp_data <- dataImputation(traj, id_field = TRUE, method = 2,
+imp_data <- data_imputation(traj, id_field = TRUE, method = 2,
 replace_with = 1,
 fill_zeros = FALSE, verbose=FALSE)
 
