+++ man/alpha_label.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/alphaLabel.R
-\name{alphaLabel}
-\alias{alphaLabel}
+% Please edit documentation in R/alpha_label.R
+\name{alpha_label}
+\alias{alpha_label}
 \title{Numerics ids to alphabetical ids}
 \usage{
-alphaLabel(x)
+alpha_label(x)
 }
 \arguments{
 \item{x}{A vector of numeric ids}
@@ -18,7 +18,7 @@ to alphabetic ids
 }
 \details{
 Given a vector of numeric cluster ids,
-`alphaLabel` converts each id to its corresponding alphabets.
+`alpha_label` converts each id to its corresponding alphabets.
 It combines alphabets for ids greater than 26.
 }
 \examples{
@@ -35,7 +35,7 @@ numeric_Labels <- kml::affectIndivC(TO1Risk, center)
 mode(numeric_Labels)
 
 #transform numeric cluster labels to alphabets
-alphab_Labels <- alphaLabel(numeric_Labels)
+alphab_Labels <- alpha_label(numeric_Labels)
 
 mode(alphab_Labels)
 
