+++ man/elbow_point.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/elbowPoint.R
-\name{elbowPoint}
-\alias{elbowPoint}
+% Please edit documentation in R/elbow_point.R
+\name{elbow_point}
+\alias{elbow_point}
 \title{Determine the elbow point on a curve}
 \usage{
-elbowPoint(x, y)
+elbow_point(x, y)
 }
 \arguments{
 \item{x}{vector of x coordinates of points on the curve}
@@ -27,10 +27,10 @@ elbow point (credit: 'github.com/agentlans')
 
 data(TO1Risk)
 
-output <- akClustr(TO1Risk, id_field = FALSE,
+output <- akclustr(TO1Risk, id_field = FALSE,
                   method = "linear", k = c(3,20),
                   crit='Silhouette',
-                  verbose = TRUE, quality_Plot=TRUE)
+                  verbose = TRUE, quality_plot=TRUE)
 
 print(output$optimal_k)
 
