### 2020-11-18 cc62f93
update manuals
