+++ README.Rmd
@@ -1,6 +1,6 @@
 ---
-output: github_document
-#output: html_document
+#output: github_document
+output: html_document
 ---
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
