+++ R/print_akStats.R
@@ -25,9 +25,9 @@
 #' This result can be used to further categorize the groups into
 #' 'classes'. For example, groups that fall within the \code{1st}
 #' quartile may be classified as 'Stable' groups (Adepeju et al. 2020).
-#' @param show_plots [TRUE or FALSE] To display cluster plots.
-#' Line plots are displayed by default. Please, see
-#' \code{plot_akstats} function for more plot options.
+#' @param show_plots [TRUE or FALSE] Provides the trajectory group
+#' plot. Please, see \code{plot_akstats} function for more
+#' plot options.
 #' Defaults \code{FALSE}
 #' @s3method print_akstats
 #' @examples
