+++ README.Rmd
@@ -244,7 +244,7 @@ prpties$change_stats
 (See the package vignette for more details about the interpretation of the tables)
 
 
-## (iii) `plot_akstats` function:
+### Plotting the groups
 
 The above printouts represent the properties (i.e. the descriptive and change properties) of the clusters. Note that the `print_akstats` function also include `show_plots` argument to show the plot of group trajectories. However, the `plot_akstats` is provided to provide a more comprehensive performance plots of the groups. Examples: 
 
