### 2020-11-19 d125fa5
update clustering algo.
