### 2020-11-19 2348217
removed manuals
