+++ man/plot.numberNeededSim.Rd
@@ -1,22 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/plot.numberNeededSim.R
-\name{plot.numberNeededSim}
-\alias{plot.numberNeededSim}
-\title{Plot of cluster groups.}
-\usage{
-\method{plot}{numberNeededSim}(ak_object, ...)
-}
-\arguments{
-\item{ak_object}{An output of \code{\link{akclustr}} function.
-The object contains individual trajectories and their cluster
-solution(s) at the specified values of \code{k}. Also, includes
-the optimal value of \code{k} based on the criterion specified.
-at (different) values of \code{k} the \code{traj}.}
-
-\item{...}{Not used at the moment.}
-}
-\description{
-Takes the 'ak_object' from the
-\code{'akclustr'} as input and produce either the 'line' plot
-or 'stacked' histogram.
-}
