+++ man/print.numberNeededSim.Rd
@@ -1,24 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/print.numberNeededSim.R
-\name{print.numberNeededSim}
-\alias{print.numberNeededSim}
-\title{Descriptive (Change) statistics}
-\usage{
-\method{print}{numberNeededSim}(ak_object, ...)
-}
-\arguments{
-\item{ak_object}{An output of \code{\link{akclustr}} function.
-The object contains individual trajectories and their cluster
-solution(s) at the specified values of \code{k}. Also, includes
-the optimal value of \code{k} based on the criterion specified.
-at (different) values of \code{k} the \code{traj}.}
-
-\item{...}{Not used at the moment.}
-}
-\description{
-This function perform two tasks:
-(i) it generate the descriptive and change statistics
-of groups, particularly suited for the outputs form
-the \code{\link{akclustr}} function, and
-(ii) generates the plots of the groups (performances).
-}
