+++ NAMESPACE
@@ -1,14 +1,11 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(plot,numberNeededSim)
 S3method(plot_akstats,default)
-S3method(print,numberNeededSim)
 S3method(print_akstats,default)
 export(akclustr)
 export(alpha_label)
 export(data_imputation)
 export(elbow_point)
-export(numberNeededSim)
 export(outlier_detect)
 export(plot_akstats)
 export(print_akstats)
