+++ R/plot.numberNeededSim.R
@@ -1,214 +1,0 @@
-#' @title Plot of cluster groups.
-#' @description Takes the 'ak_object' from the
-#' \code{'akclustr'} as input and produce either the 'line' plot
-#' or 'stacked' histogram.
-#' @param ak_object An output of \code{\link{akclustr}} function.
-#' The object contains individual trajectories and their cluster
-#' solution(s) at the specified values of \code{k}. Also, includes
-#' the optimal value of \code{k} based on the criterion specified.
-#' at (different) values of \code{k} the \code{traj}.
-#' @param ... Not used at the moment.
-#' @importFrom reshape2 melt
-#' @importFrom stats quantile
-#' @importFrom utils flush.console
-#' @importFrom grDevices dev.new
-#' @importFrom ggplot2 stat_summary scale_colour_brewer theme_light
-#' @export
-plot.numberNeededSim <- function(ak_object, ...){
-
-  k <- 3
-  reference <- 1
-  n_quant <- 4
-  type <- "lines"
-  y_scaling <- "fixed"
-
-  #first testing that correct values of k is specified.
-  #get all values of k..
-  all_K <- as.vector(unlist(lapply(ak_object$solutions, attributes)))
-
-  if(!k %in% all_K){
-    stop(paste("*----k =", k, "is not applicable!. Print the",
-               "'akobject' to see allowed k-values----*", sep=" "))
-  }
-
-  # check object type
-  # if(class(ak_object)[1] != "akobject"){
-  #   stop("*----Object not right type!! 'akclustr' object required!----*")
-  # }
-
-  #test data type/or class
-
-  #extract variables
-  traj <- ak_object$traj
-  clustr <- as.vector(ak_object$solutions[[k-2]])
-  id_field <- ak_object$id_field
-
-  #testing that data and clusters have equal number of elements
-  if(length(clustr)!=nrow(traj)){
-    stop("*----Unequal number of clusters elements and trajectories----*")
-  }
-
-  #joining the data with clusters
-  clustr <- data.frame(cbind(traj, clusters=clustr))
-
-  dat <- traj #back up traj
-
-  n_quant <- round(n_quant, digits = 0)
-
-  if(n_quant < 2 | n_quant > 10){
-    stop(paste("*----Please, enter an integer between 2",
-               "and 10 for the 'n_quant' argument'!!!----*", sep=" "))
-  }
-
-  #test id_field is true
-  if(id_field==TRUE){
-    dat <- dat[,2:ncol(dat)]
-    n_CL <- colnames(clustr)[1]
-    col_names <- as.vector(clustr[,1])
-
-    #test if id field  is unique
-    if(!length(col_names)==length(unique(col_names))){
-      stop(paste("(: The 'id_field' is not a unique field.",
-                 "Function terminated!!! :)", sep=" "))
-    }
-  }
-
-  #test if id_field is excluded for traj
-  if(id_field==FALSE){
-    clustr <- cbind(seq_len(nrow(clustr)), clustr)
-  }
-
-  #collect cluster list
-  clusters <- as.vector(clustr[,ncol(clustr)])
-
-  data_subset <- clustr[,seq_len((ncol(clustr))-1)]
-
-  data_subset <- as.data.frame(data_subset)
-
-  colnames(data_subset) <- c("code", seq_len((ncol(data_subset))-1))
-
-  data.subset.melted <- suppressWarnings(melt(data_subset, id="code"))
-
-  #append cluster list with traj
-  data.subset.melted <- cbind(data.subset.melted,
-                              rep(clusters, ncol(data_subset)-1))
-  colnames(data.subset.melted) <- c("id","Year","value", "clusters")
-
-  #----------------------------------------------------
[…]
