### 2020-11-24 #2379 @MAnalytics
@karthik Please, may I ask that you re-generate the .pdf as I modified the paper slightly. Specifically, in line with the revisions made, I edited this bit.....

"The statPrint function draws from the ggplot2 library (Wickham, 2016) in order to visualise
the resulting clusters in either a line or an areal-stacked graph format, alongside descriptive
cluster statistics."

Edited as:

"The package includes the `print_akstats` and `plot_akstats` functions to generate the properties of clusters and visualize the clusters, respectively. ....".

Thanks.
