### 2020-11-26 #2379 @karthik
@MAnalytics sure thing. You can also run the command above to generate a new pdf.
