### 2020-11-27 #2379 @karthik
@MAnalytics Yes, please confirm the version number of your package. Then you'll have to archive the software on Zenodo and share the DOI here.
