### 2020-11-27 #2379 @MAnalytics
> @MAnalytics Yes, please confirm the version number of your package. Then you'll have to archive the software on Zenodo and share the DOI here.

@karthik I will like to confirm the version number of the package as Version 0.2.5. The 'DESCRIPTION' file has been modified accordingly. Thanks.
