### 2020-11-28 #2379 @MAnalytics
@karthik This is to inform you that I have archived the software on Zenovo, and below is the version and the DOI:

Akmedoids v2.0.0
DOI: 10.5281/zenodo.4294266

Thanks
