### 2020-11-30 #2379 @karthik
@MAnalytics Can you please edit your Zenodo submission and remove your GitHub username from the title. Just `akmedoids v2.0.0`. Then please update the DOI here.
