### 2020-12-02 #2379 @kyleniemeyer
Hello @MAnalytics, I am the editor-in-chief on duty this week, doing some final checks of your submission.

One of the [required elements of JOSS papers](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) is an explicit Statement of Need section. I think you could just rename your Summary to Statement of Need. Once you've made that change, I can publish the article—no need to archive the repository again.
