### 2020-12-02 #2379 @kyleniemeyer
@MAnalytics ok, the JOSS publication is ready to go. What is the status of the associated AAS article at this point?
