### 2020-12-02 #2379 @MAnalytics
@kyleniemeyer Sorry, there isn't any associated AAS article. I mistakenly included the "aas-journal of open source software" line in the .md file header. This has now been corrected. Thanks.
